split standard/template/xml/json output code in different files, fix #18
