+++ Makefile.noauto
@@ -10,7 +10,7 @@ CFLAGS = -DDEBUG=1 -DENABLE_DUMP=1
 ## NOTE: if you get errors when linking qstat (missing symbols or
 ## libraries), then modify LDFLAGS or LDLIBS
 
-SRC = utils.c xform.c config.c debug.c hcache.c md5.c qserver.c qstat.c template.c ut2004.c a2s.c packet_manip.c gs3.c gs2.c gps.c ts2.c doom3.c tm.c haze.c wic.c ottd.c fl.c tee.c ts3.c bfbc2.c ventrilo.c cube2.c mumble.c terraria.c crysis.c dirtybomb.c starmade.c farmsim.c ksp.c
+SRC = utils.c xform.c config.c debug.c hcache.c md5.c qserver.c qstat.c display_standard.c display_template.c display_raw.c display_xml.c display_json.c ut2004.c a2s.c packet_manip.c gs3.c gs2.c gps.c ts2.c doom3.c tm.c haze.c wic.c ottd.c fl.c tee.c ts3.c bfbc2.c ventrilo.c cube2.c mumble.c terraria.c crysis.c dirtybomb.c starmade.c farmsim.c ksp.c
 OBJ = $(SRC:.c=.obj)
 O   = $(SRC:.c=.o)
