+++ qstat.c
@@ -7,7 +7,6 @@
  * Thanks to Per Hammer for the OS/2 patches (per@mindbend.demon.co.uk)
  * Thanks to John Ross Hunt for the OpenVMS Alpha patches (bigboote@ais.net)
  * Thanks to Scott MacFiggen for the quicksort code (smf@webmethods.com)
- * Thanks to Simon Garner for the XML patch (sgarner@gameplanet.co.nz)
  * Thanks to Bob Marriott for the Ghost Recon code (bmarriott@speakeasy.net)
  *
  * Inspired by QuakePing by Len Norton
@@ -34,11 +33,11 @@
 #include <assert.h>
 
 #define QUERY_PACKETS
+
 #include "qstat.h"
 #include "packet_manip.h"
 #include "config.h"
 #include "xform.h"
-
 
 #ifndef _WIN32
 	#include <signal.h>
@@ -162,21 +161,10 @@ int up_servers_only = 0;
 int no_full_servers = 0;
 int no_empty_servers = 0;
 int no_header_display = 0;
-int raw_display = 0;
-char *raw_delimiter = "\t";
-char *multi_delimiter = "|";
 int player_address = 0;
 int max_simultaneous = MAXFD_DEFAULT;
 int sendinterval = 5;
-extern int xform_names;
-extern int xform_strip_unprintable;
-extern int xform_hex_player_names;
-extern int xform_hex_server_names;
-extern int xform_strip_carets;
-extern int xform_html_names;
-extern int html_mode;
-int raw_arg = 0;
-int show_game_in_raw = 0;
+
 int progress = 0;
 int num_servers_total = 0;
 int num_players_total = 0;
@@ -184,8 +172,10 @@ int max_players_total = 0;
 int num_servers_returned = 0;
 int num_servers_timed_out = 0;
 int num_servers_down = 0;
+
+char *multi_delimiter = "|";
+
 server_type *default_server_type = NULL;
-FILE *OF; /* output file */
 unsigned int source_ip = INADDR_ANY;
 unsigned short source_port_low = 0;
 unsigned short source_port_high = 0;
@@ -193,36 +183,12 @@ unsigned short source_port = 0;
 int show_game_port = 0;
 int no_port_offset = 0;
 
-#define ENCODING_LATIN_1 1
-#define ENCODING_UTF_8  8
-#define UTF8BYTESWAPNOTACHAR	0xFFFE
-#define UTF8NOTACHAR	0xFFFF
-#define UTF8MAXFROMUCS4	0x10FFFF
-
-int output_bom = 0;
-int xml_display = 0;
-int xml_encoding = ENCODING_LATIN_1;
-int json_display = 0;
-int json_printed = 0;
-
 #define SUPPORTED_SERVER_SORT	"pgihn"
 #define SUPPORTED_PLAYER_SORT	"PFTNS"
 #define SUPPORTED_SORT_KEYS	"l" SUPPORTED_SERVER_SORT SUPPORTED_PLAYER_SORT
 char sort_keys[32];
 int player_sort = 0;
 int server_sort = 0;
-
-void quicksort(void **array, int i, int j, int(*compare)(void *, void*));
-int qpartition(void **array, int i, int j, int(*compare)(void *, void*));
-void sort_servers(struct qserver **array, int size);
-void sort_players(struct qserver *server);
-int server_compare(struct qserver *one, struct qserver *two);
-int player_compare(struct player *one, struct player *two);
-int type_option_compare(server_type *one, server_type *two);
-int type_string_compare(server_type *one, server_type *two);
-int process_func_ret( struct qserver *server, int ret );
-int connection_inprogress();
-void clear_socketerror();
 
 int show_errors = 0;
 static int noserverdups = 1;
@@ -240,10 +206,11 @@ int time_format = -1;
 
 struct qserver *servers = NULL;
 struct qserver **last_server = &servers;
-struct qserver **connmap = NULL;
-int max_connmap;
 struct qserver *last_server_bind = NULL;
 struct qserver *first_server_bind = NULL;
+
+struct qserver **connmap = NULL;
+int max_connmap;
 int connected = 0;
 time_t run_timeout = 0;
 time_t start_time;
@@ -254,7 +221,6 @@ static unsigned num_servers; /* current number of servers in memory */
 static struct qserver **server_hash[ADDRESS_HASH_LENGTH];
 static unsigned int server_hash_len[ADDRESS_HASH_LENGTH];
 static void free_server_hash();
-static void xml_display_player_info_info(struct player *player);
 
 char *DOWN = "DOWN";
 char *SYSERROR = "SYSERROR";
@@ -266,6 +232,7 @@ char *BFRIS_SERVER_NAME = "BFRIS Server";
 char *GAMESPY_MASTER_NAME = "Gamespy Master";
 
 int display_prefix = 0;
+
 char *current_filename;
 int current_fileline;
 
@@ -279,14 +246,18 @@ static int decode_q3master_packet(struct qserver *server, char *ikt, int pktlen)
 
 char *ut2003_strdup(const char *string, const char *end, char **next);
 
-void free_server(struct qserver *server);
-void free_player(struct player *player);
-void free_rule(struct rule *rule);
-void standard_display_server(struct qserver *server);
+void set_non_blocking(int fd);
+int set_fds(fd_set *fds);
+void get_next_timeout(struct timeval *timeout);
 
-/* MODIFY HERE
- * Change these functions to display however you want
+void set_file_descriptors();
+int wait_for_file_descriptors(struct timeval *timeout);
+struct qserver *get_next_ready_server();
+
+/*
+ * Functions
  */
+
 void display_server(struct qserver *server)
 {
 	if (player_sort)
@@ -316,3256 +287,6 @@ void display_server(struct qserver *server)
 	}
 
 	free_server(server);
-}
-
-void standard_display_server(struct qserver *server)
-{
-	char prefix[64];
-	if (display_prefix)
-	{
-		sprintf(prefix, "%-4s ", server->type->type_prefix);
-	}
-	else
-	{
-		prefix[0] = '\0';
-	}
-
-	if (server->server_name == DOWN || server->server_name == SYSERROR)
-	{
-		if (!up_servers_only)
-		{
-			xform_printf(OF, "%s%-16s %10s\n", prefix, (hostname_lookup) ? server->host_name: server->arg, server->server_name);
-		}
-		return ;
-	}
-
-	if (server->server_name == TIMEOUT)
-	{
-		if (server->flags &FLAG_BROADCAST && server->n_servers)
-		{
-			xform_printf(OF, "%s%-16s %d servers\n", prefix, server->arg, server->n_servers);
-		}
-		else if (!up_servers_only)
-		{
-			xform_printf(OF, "%s%-16s no response\n", prefix, (hostname_lookup) ? server->host_name: server->arg);
-		}
-		return ;
-	}
-
-	if (server->type->master)
-	{
-		display_qwmaster(server);
-		return ;
-	}
-
-	if (no_full_servers && server->num_players >= server->max_players)
-	{
-		return ;
-	}
-
-	if (no_empty_servers && server->num_players == 0)
-	{
-		return ;
-	}
-
-	if (server->error != NULL)
-	{
-		xform_printf(OF, "%s%-21s ERROR <%s>\n", prefix, (hostname_lookup) ? server->host_name: server->arg, server->error);
-		return ;
-	}
-
-	if (new_style)
-	{
-		char *game = get_qw_game(server);
-		int map_name_width = 8, game_width = 0;
-		switch (server->type->id)
-		{
-			case QW_SERVER:
-			case Q2_SERVER:
-			case Q3_SERVER:
-				game_width = 9;
-				break;
-			case TRIBES2_SERVER:
-				map_name_width = 14;
-				game_width = 8;
-				break;
-			case GHOSTRECON_SERVER:
-				map_name_width = 15;
-				game_width = 15;
-				break;
-			case HL_SERVER:
-				map_name_width = 12;
-				break;
-			default:
-				break;
-		}
-		xform_printf(OF, 
-			"%s%-21s %2d/%-2d %2d/%-2d %*s %6d / %1d  %*s %s\n",
-			prefix,
-			(hostname_lookup) ? server->host_name: server->arg,
-			server->num_players,
-			server->max_players,
-			server->num_spectators,
-			server->max_spectators,
-			map_name_width,
-			(server->map_name) ? xform_name(server->map_name, server): "?",
-			server->n_requests ? server->ping_total / server->n_requests: 999,
-			server->n_retries,
-			game_width,
-			game,
-			xform_name(server->server_name, server)
-		);
-		if (get_server_rules && NULL != server->type->display_rule_func )
-		{
-			server->type->display_rule_func(server);
-		}
-		if (get_player_info && NULL != server->type->display_player_func )
-		{
-			server->type->display_player_func(server);
-		}
-	}
-	else
-	{
-		char name[512];
-		sprintf(name, "\"%s\"", server->server_name);
-		xform_printf(OF, 
-			"%-16s %10s map %s at %22s %d/%d players %d ms\n",
-			(hostname_lookup) ? server->host_name: server->arg,
-			name,
-			server->map_name,
-			server->address,
-			server->num_players,
-			server->max_players,
-			server->n_requests ? server->ping_total / server->n_requests: 999
-		);
-	}
-}
-
-void display_qwmaster(struct qserver *server)
-{
-	char *prefix;
-	prefix = server->type->type_prefix;
-
-	if (server->error != NULL)
-	{
-		xform_printf(OF, 
-			"%s %-17s ERROR <%s>\n",
-			prefix,
-			(hostname_lookup) ? server->host_name: server->arg,
-			server->error
-		);
-	}
-	else
-	{
-		xform_printf(OF, 
-			"%s %-17s %d servers %6d / %1d\n",
-			prefix,
-			(hostname_lookup) ? server->host_name: server->arg,
-			server->n_servers,
-			server->n_requests ? server->ping_total / server->n_requests: 999,
-			server->n_retries
-		);
-	}
-}
-
-void display_header()
-{
-	if (!no_header_display)
-	{
-		xform_printf(OF, "%-16s %8s %8s %15s    %s\n", "ADDRESS", "PLAYERS", "MAP", "RESPONSE TIME", "NAME");
-	}
-}
-
-void display_server_rules(struct qserver *server)
-{
-	struct rule *rule;
-	int printed = 0;
-	rule = server->rules;
-	for (; rule != NULL; rule = rule->next)
-	{
-		if ((server->type->id != Q_SERVER && server->type->id != H2_SERVER) || !is_default_rule(rule))
-		{
-			xform_printf(OF, "%c%s=%s", (printed) ? ',' : '\t', rule->name, rule->value);
-			printed++;
-		}
-	}
-	if (printed)
-	{
-		fputs("\n", OF);
-	}
-}
-
-void display_q_player_info(struct qserver *server)
-{
-	char fmt[128];
-	struct player *player;
-
-	strcpy(fmt, "\t#%-2d %3d frags %9s ");
-
-	if (color_names)
-	{
-		strcat(fmt, "%9s:%-9s ");
-	}
-	else
-	{
-		strcat(fmt, "%2s:%-2s ");
-	}
-	if (player_address)
-	{
-		strcat(fmt, "%22s ");
-	}
-	else
-	{
-		strcat(fmt, "%s");
-	}
-	strcat(fmt, "%s\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt,
-			player->number,
-			player->frags,
-			play_time(player->connect_time, 1),
-			quake_color(player->shirt_color),
-			quake_color(player->pants_color),
-			(player_address) ? player->address: "",
-			xform_name(player->name, server)
-		);
-	}
-}
-
-void display_qw_player_info(struct qserver *server)
-{
-	char fmt[128];
-	struct player *player;
-
-	strcpy(fmt, "\t#%-6d %5d frags %6s@%-5s %8s");
-
-	if (color_names)
-	{
-		strcat(fmt, "%9s:%-9s ");
-	}
-	else
-	{
-		strcat(fmt, "%2s:%-2s ");
-	}
-	strcat(fmt, "%12s %s\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt,
-			player->number,
-			player->frags,
-			play_time(player->connect_time, 0),
-			ping_time(player->ping),
-			player->skin ? player->skin: "",
-			quake_color(player->shirt_color),
-			quake_color(player->pants_color),
-			xform_name(player->name, server),
-			xform_name(player->team_name, server)
-		);
-	}
-}
-
-void display_q2_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (server->flags &FLAG_PLAYER_TEAMS)
-		{
-			xform_printf(OF, "\t%3d frags team#%d %8s  %s\n", player->frags, player->team, ping_time(player->ping), xform_name(player->name, server));
-		}
-		else
-		{
-			xform_printf(OF, "\t%3d frags %8s  %s\n", player->frags, ping_time(player->ping), xform_name(player->name, server));
-		}
-	}
-}
-
-
-
-void display_unreal_player_info(struct qserver *server)
-{
-	struct player *player;
-	static const char *fmt_team_number = "\t%3d frags team#%-3d %7s %s\n";
-	static const char *fmt_team_name = "\t%3d frags %8s %7s %s\n";
-	static const char *fmt_no_team = "\t%3d frags %8s  %s\n";
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (server->flags &FLAG_PLAYER_TEAMS)
-		{
-			// we use (player->score) ? player->score : player->frags,
-			// so we get details from halo
-			if (player->team_name)
-			{
-				xform_printf(OF, fmt_team_name,
-					(player->score && NA_INT != player->score) ? player->score: player->frags,
-					player->team_name,
-					ping_time(player->ping),
-					xform_name(player->name, server)
-				);
-			}
-			else
-			{
-				xform_printf(OF, fmt_team_number,
-					(player->score && NA_INT != player->score) ? player->score: player->frags,
-					player->team,
-					ping_time(player->ping),
-					xform_name(player->name, server)
-				);
-			}
-		}
-		else
-		{
-			xform_printf(OF, fmt_no_team, player->frags, ping_time(player->ping), xform_name(player->name, server));
-		}
-	}
-}
-
-void display_shogo_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t%3d frags %8s %s\n", player->frags, ping_time(player->ping), xform_name(player->name, server));
-	}
-}
-
-void display_halflife_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t%3d frags %8s %s\n", player->frags, play_time(player->connect_time, 1), xform_name(player->name, server));
-	}
-}
-
-void display_fl_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t%3d frags %8s %8s %s\n", player->frags, ping_time(player->ping), play_time(player->connect_time, 1), xform_name(player->name, server));
-	}
-}
-
-void display_tribes_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t%4d score team#%d %8s %s\n", player->frags, player->team, ping_time(player->ping), xform_name(player->name, server)
-	);
-	}
-}
-
-void display_tribes2_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\tscore %4d %14s %s\n",
-			player->frags,
-			player->team_name ? player->team_name: (player->number == TRIBES_TEAM ?	"TEAM" : "?"),
-			xform_name(player->name, server)
-		);
-	}
-}
-
-void display_bfris_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\ttid: %d, ship: %d, team: %s, ping: %d, score: %d, kills: %d, name: %s\n",
-			player->number,
-			player->ship,
-			player->team_name,
-			player->ping,
-			player->score,
-			player->frags,
-			xform_name(player->name, server)
-		);
-	}
-}
-
-void display_descent3_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t%3d frags %3d deaths team#%-3d %7s %s\n",
-			player->frags,
-			player->deaths,
-			player->team,
-			ping_time(player->ping),
-			xform_name(player->name, server)
-		);
-	}
-}
-
-void display_ghostrecon_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\tdead=%3d team#%-3d %s\n", player->deaths, player->team, xform_name(player->name, server));
-	}
-}
-
-void display_eye_player_info(struct qserver *server)
-{
-	struct player *player;
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (player->team_name)
-		{
-			xform_printf(OF, "\tscore %4d %6s team %12s %s\n",
-				player->score,
-				ping_time(player->ping),
-				player->team_name,
-				xform_name(player->name,server)
-			);
-		}
-		else
-		{
-			xform_printf(OF, "\tscore %4d %6s team#%d %s\n",
-				player->score,
-				ping_time(player->ping),
-				player->team,
-				xform_name(player->name,server)
-			);
-		}
-	}
-}
-
-int calculate_armyops_score(struct player *player)
-{
-	/* Calculates a player's score for ArmyOps from the basic components */
-
-	int score = 0;
-	int kill_score = 0;
-	struct info *info;
-
-	for (info = player->info; info; info = info->next)
-	{
-		if (0 == strcmp(info->name, "leader") || 0 == strcmp(info->name, "goal") || 0 == strcmp(info->name, "roe"))
-		{
-			score += atoi(info->value);
-		}
-		else if (0 == strcmp(info->name, "kia") || 0 == strcmp(info->name, "enemy"))
-		{
-			kill_score += atoi(info->value);
-		}
-	}
-
-	if (kill_score > 0)
-	{
-		score += kill_score;
-	}
-
-	return score;
-}
-
-void display_gs2_player_info(struct qserver *server)
-{
-	struct player *player;
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (player->team_name)
-		{
-			xform_printf(OF, "\tscore %4d %6s team %12s %s\n",
-				player->score,
-				ping_time(player->ping),
-				player->team_name,
-				xform_name(player->name,server)
-			);
-		}
-		else
-		{
-			xform_printf(OF, "\tscore %4d %6s team#%d %s\n",
-				player->score,
-				ping_time(player->ping),
-				player->team,
-				xform_name(player->name, server)
-			);
-		}
-	}
-}
-
-void display_armyops_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		player->score = calculate_armyops_score(player);
-	}
-
-	display_gs2_player_info(server);
-}
-
-void display_ts2_player_info(struct qserver *server)
-{
-	struct player *player;
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t%6s %s\n", ping_time(player->ping), xform_name(player->name, server));
-	}
-}
-
-void display_ts3_player_info(struct qserver *server)
-{
-	struct player *player;
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t%s\n", xform_name(player->name, server));
-	}
-}
-
-void display_starmade_player_info(struct qserver *server)
-{
-	struct player *player;
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t%s\n", xform_name(player->name, server));
-	}
-}
-
-void display_bfbc2_player_info(struct qserver *server)
-{
-	struct player *player;
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t%s\n", xform_name(player->name, server));
-	}
-}
-
-void display_wic_player_info(struct qserver *server)
-{
-	struct player *player;
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t#%-4d score %4d team %12s role %12s %s\n",
-			player->number,
-			player->score,
-			player->team_name,
-			player->tribe_tag ? player->tribe_tag : "",
-			xform_name(player->name, server)
-		);
-	}
-}
-
-void display_ventrilo_player_info(struct qserver *server)
-{
-	struct player *player;
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t# %d ping %s time %d cid %i ch %s name %s\n",
-			player->number,
-			ping_time(player->ping),
-			player->connect_time,
-			player->team,
-			player->team_name,
-			xform_name(player->name, server)
-		);
-	}
-}
-
-void display_tm_player_info(struct qserver *server)
-{
-	struct player *player;
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t%6s %s\n", ping_time(player->ping), xform_name(player->name, server));
-	}
-}
-
-void display_doom3_player_info(struct qserver *server)
-{
-	struct player *player;
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (player->tribe_tag)
-		{
-			xform_printf(OF, "\t#%-4d score %4d %6s team %12s %s\n",
-				player->number,
-				player->score,
-				ping_time(player->ping),
-				player->tribe_tag,
-				xform_name(player->name, server)
-			);
-		}
-		else
-		{
-			xform_printf(OF, "\t#%-4d score %4d %6s team#%d %s\n",
-				player->number,
-				player->score,
-				ping_time(player->ping),
-				player->team,
-				xform_name(player->name, server)
-			);
-		}
-	}
-}
-
-void display_ravenshield_player_info(struct qserver *server)
-{
-	struct player *player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t%3d frags %8s %s\n", player->frags, play_time(player->connect_time, 1), xform_name(player->name, server));
-	}
-}
-
-
-void display_savage_player_info(struct qserver *server)
-{
-	struct player *player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t%3d frags %8s %s\n", player->frags, play_time(player->connect_time, 1), xform_name(player->name, server));
-	}
-}
-
-
-void display_farcry_player_info(struct qserver *server)
-{
-	struct player *player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t%3d frags %8s %s\n", player->frags, play_time(player->connect_time, 1), xform_name(player->name, server));
-	}
-}
-
-void display_tee_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t%4d score %s\n", player->score, xform_name(player->name, server));
-	}
-}
-
-char *get_qw_game(struct qserver *server)
-{
-	struct rule *rule;
-	char *game_rule = server->type->game_rule;
-	if (game_rule == NULL || *game_rule == '\0')
-	{
-		return "";
-	} rule = server->rules;
-	for (; rule != NULL; rule = rule->next)
-	{
-		if (strcmp(rule->name, game_rule) == 0)
-		{
-			if (server->type->id == Q3_SERVER && strcmp(rule->value, "baseq3") == 0)
-			{
-				return "";
-			}
-			return rule->value;
-		}
-	}
-	rule = server->rules;
-	for (; rule != NULL; rule = rule->next)
-	{
-		if (0 == strcmp(rule->name, "game") || 0 == strcmp(rule->name, "fs_game"))
-		{
-			return rule->value;
-		}
-	}
-	return "";
-}
-
-/* Raw output for web master types
- */
-
-#define RD raw_delimiter
-
-void raw_display_server(struct qserver *server)
-{
-	char *prefix;
-	int ping_time;
-	prefix = server->type->type_prefix;
-
-	if (server->n_requests)
-	{
-		ping_time = server->ping_total / server->n_requests;
-	}
-	else
-	{
-		ping_time = 999;
-	}
-
-	if (server->server_name == DOWN || server->server_name == SYSERROR)
-	{
-		if (!up_servers_only)
-		{
-			xform_printf(OF, "%s""%.*s%.*s""%s%s""%s%s\n\n",
-				prefix,
-				raw_arg, RD,
-				raw_arg,
-				server->arg, RD,
-				(hostname_lookup) ? server->host_name : server->arg, RD,
-				server->server_name
-			);
-		}
-		return ;
-	}
-
-	if (server->server_name == TIMEOUT)
-	{
-		if (server->flags &FLAG_BROADCAST && server->n_servers)
-		{
-			xform_printf(OF, "%s""%.*s%.*s""%s%s""%s%d\n", prefix, raw_arg, RD, raw_arg, server->arg, RD, server->arg, RD, server->n_servers);
-		}
-		else if (!up_servers_only)
-		{
-			xform_printf(OF, "%s""%.*s%.*s""%s%s""%s%s\n\n", prefix, raw_arg, RD, raw_arg, server->arg, RD, (hostname_lookup) ? server->host_name : server->arg, RD, TIMEOUT);
-		}
-		return ;
-	}
-
-	if (server->error != NULL)
-	{
-		xform_printf(OF, "%s""%.*s%.*s""%s%s""%s%s""%s%s",
-			prefix,
-			raw_arg, RD,
-			raw_arg,
-			server->arg, RD,
-			(hostname_lookup) ? server->host_name: server->arg, RD,
-			"ERROR", RD,
-			server->error
-		);
-	}
-	else if (server->type->flags &TF_RAW_STYLE_QUAKE)
-	{
-		xform_printf(OF, "%s""%.*s%.*s""%s%s""%s%s""%s%s""%s%d""%s%s""%s%d""%s%d""%s%d""%s%d""%s%d""%s%d""%s%s",
-			prefix,
-			raw_arg, RD,
-			raw_arg,
-			server->arg, RD,
-			(hostname_lookup) ? server->host_name: server->arg, RD,
-			xform_name(server->server_name, server), RD,
-			server->address, RD,
-			server->protocol_version, RD,
-			server->map_name, RD,
-			server->max_players, RD,
-			server->num_players, RD,
-			server->max_spectators, RD,
-			server->num_spectators, RD,
-			ping_time, RD,
-			server->n_retries,
-			show_game_in_raw ? RD : "",
-			show_game_in_raw ? get_qw_game(server): ""
-		);
-	}
-	else if (server->type->flags &TF_RAW_STYLE_TRIBES)
-	{
-		xform_printf(OF, "%s""%.*s%.*s""%s%s""%s%s""%s%s""%s%d""%s%d",
-			prefix,
-			raw_arg, RD,
-			raw_arg,
-			server->arg, RD,
-			(hostname_lookup) ?	server->host_name: server->arg, RD,
-			xform_name(server->server_name, server), RD,
-			(server->map_name) ? server->map_name: "?", RD,
-			server->num_players, RD,
-			server->max_players
-		);
-	}
-	else if (server->type->flags &TF_RAW_STYLE_GHOSTRECON)
-	{
-		xform_printf(OF, "%s""%.*s%.*s""%s%s""%s%s""%s%s""%s%d""%s%d",
-			prefix,
-			raw_arg, RD,
-			raw_arg,
-			server->arg, RD,
-			(hostname_lookup) ?	server->host_name: server->arg, RD,
-			xform_name(server->server_name,	server), RD,
-			(server->map_name) ? server->map_name: "?", RD,
-			server->num_players, RD,
-			server->max_players
-		);
-	}
-	else if (server->type->master)
-	{
-		xform_printf(OF, "%s""%.*s%.*s""%s%s""%s%d",
-			prefix,
-			raw_arg, RD,
-			raw_arg,
-			server->arg, RD,
-			(hostname_lookup) ? server->host_name: server->arg, RD,
-			server->n_servers
-		);
-	}
-	else
-	{
-		xform_printf(OF, "%s""%.*s%.*s""%s%s""%s%s""%s%s""%s%d""%s%d""%s%d""%s%d""%s%s",
-			prefix,
-			raw_arg, RD,
-			raw_arg,
-			server->arg, RD,
-			(hostname_lookup) ? server->host_name: server->arg, RD,
-			xform_name(server->server_name,	server), RD,
-			(server->map_name) ? xform_name(server->map_name, server): "?", RD,
-			server->max_players, RD,
-			server->num_players, RD,
-			ping_time, RD,
-			server->n_retries,
-			show_game_in_raw ? RD : "",
-			show_game_in_raw ? get_qw_game(server): ""
-		);
-	}
-	fputs("\n", OF);
-
-	if (server->type->master || server->error != NULL)
-	{
-		fputs("\n", OF);
-		return ;
-	}
-
-	if (get_server_rules && NULL != server->type->display_raw_rule_func )
-	{
-		server->type->display_raw_rule_func(server);
-	}
-	if (get_player_info && NULL != server->type->display_raw_player_func)
-	{
-		server->type->display_raw_player_func(server);
-	}
-	fputs("\n", OF);
-}
-
-void raw_display_server_rules(struct qserver *server)
-{
-	struct rule *rule;
-	int printed = 0;
-	rule = server->rules;
-	for (; rule != NULL; rule = rule->next)
-	{
-		if (server->type->id == TRIBES2_SERVER)
-		{
-			char *v;
-			for (v = rule->value; *v; v++)
-			if (*v == '\n')
-			{
-				*v = ' ';
-			}
-		}
-
-		xform_printf(OF, "%s%s=%s", (printed) ? RD : "", rule->name, rule->value);
-		printed++;
-	}
-	if (server->missing_rules)
-	{
-		xform_printf(OF, "%s?", (printed) ? RD : "");
-	}
-	fputs("\n", OF);
-}
-
-void raw_display_q_player_info(struct qserver *server)
-{
-	char fmt[] = "%d""%s%s""%s%s""%s%d""%s%s""%s%s""%s%s";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt,
-			player->number, RD,
-			xform_name(player->name, server), RD,
-			player->address, RD,
-			player->frags, RD,
-			play_time(player->connect_time, 1), RD,
-			quake_color(player->shirt_color), RD,
-			quake_color(player->pants_color)
-		);
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_qw_player_info(struct qserver *server)
-{
-	char fmt[128];
-	struct player *player;
-
-	strcpy(fmt, "%d""%s%s""%s%d""%s%s""%s%s""%s%s");
-	strcat(fmt, "%s%d""%s%s""%s%s");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt,
-			player->number, RD,
-			xform_name(player->name, server), RD,
-			player->frags, RD,
-			play_time(player->connect_time, 1), RD,
-			quake_color(player->shirt_color), RD,
-			quake_color(player->pants_color), RD,
-			player->ping, RD,
-			player->skin ? player->skin: "", RD,
-			player->team_name ? player->team_name: ""
-		);
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_q2_player_info(struct qserver *server)
-{
-	static const char *fmt = "%s""%s%d""%s%d";
-	static const char *fmt_team = "%s""%s%d""%s%d""%s%d";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (server->flags &FLAG_PLAYER_TEAMS)
-		{
-			xform_printf(OF, fmt_team, xform_name(player->name, server), RD, player->frags, RD, player->ping, RD, player->team);
-		}
-		else
-		{
-			xform_printf(OF, fmt, xform_name(player->name, server), RD, player->frags, RD, player->ping);
-		}
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_unreal_player_info(struct qserver *server)
-{
-	static const char *fmt = "%s""%s%d""%s%d""%s%d""%s%s""%s%s""%s%s";
-	static const char *fmt_team_name = "%s""%s%d""%s%d""%s%s""%s%s""%s%s""%s%s";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (player->team_name)
-		{
-			xform_printf(OF, fmt_team_name,
-				xform_name(player->name,server), RD,
-				player->frags, RD,
-				player->ping, RD,
-				player->team_name, RD,
-				player->skin ? player->skin: "", RD,
-				player->mesh ? player->mesh: "", RD,
-				player->face ? player->face: ""
-			);
-		}
-		else
-		{
-			xform_printf(OF, fmt,
-				xform_name(player->name, server), RD,
-				player->frags, RD,
-				player->ping, RD,
-				player->team, RD,
-				player->skin ? player->skin: "", RD,
-				player->mesh ? player->mesh: "", RD,
-				player->face ? player->face: ""
-			);
-		}
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_halflife_player_info(struct qserver *server)
-{
-	static char fmt[24] = "%s""%s%d""%s%s";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt, xform_name(player->name, server), RD, player->frags, RD, play_time(player->connect_time, 1));
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_fl_player_info(struct qserver *server)
-{
-	static char fmt[24] = "%s""%s%d""%s%s""%s%d""%s%d";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		fprintf(
-			OF, fmt,
-			xform_name(player->name, server), RD,
-			player->frags, RD,
-			play_time(player->connect_time, 1), RD,
-			player->ping, RD,
-			player->team
-		);
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_tribes_player_info(struct qserver *server)
-{
-	static char fmt[24] = "%s""%s%d""%s%d""%s%d""%s%d";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt, xform_name(player->name, server), RD, player->frags, RD, player->ping, RD, player->team, RD, player->packet_loss);
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_tribes2_player_info(struct qserver *server)
-{
-	static char fmt[] = "%s""%s%d""%s%d""%s%s""%s%s""%s%s";
-	struct player *player;
-	char *type;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		switch (player->type_flag)
-		{
-			case PLAYER_TYPE_BOT:
-				type = "Bot";
-				break;
-			case PLAYER_TYPE_ALIAS:
-				type = "Alias";
-				break;
-			default:
-				type = "";
-				break;
-		}
-		xform_printf(OF, fmt,
-			xform_name(player->name, server), RD,
-			player->frags, RD,
-			player->team, RD,
-			player->team_name ? player->team_name : "TEAM", RD,
-			type, RD,
-			player->tribe_tag ? xform_name(player->tribe_tag, server): ""
-		);
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_bfris_player_info(struct qserver *server)
-{
-	static char fmt[] = "%d""%s%d""%s%s""%s%d""%s%d""%s%d""%s%s";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt,
-			player->number, RD,
-			player->ship, RD,
-			player->team_name, RD,
-			player->ping, RD,
-			player->score, RD,
-			player->frags, RD,
-			xform_name(player->name, server)
-		);
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_descent3_player_info(struct qserver *server)
-{
-	static char fmt[] = "%s""%s%d""%s%d""%s%d""%s%d";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt, xform_name(player->name, server), RD, player->frags, RD, player->deaths, RD, player->ping, RD, player->team);
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_ghostrecon_player_info(struct qserver *server)
-{
-	static char fmt[28] = "%s""%s%d""%s%d";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt, xform_name(player->name, server), RD, player->deaths, RD, player->team);
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_eye_player_info(struct qserver *server)
-{
-	static const char *fmt = "%s""%s%d""%s%d""%s%d""%s%s""%s%s";
-	static const char *fmt_team_name = "%s""%s%d""%s%d""%s%s""%s%s""%s%s";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (player->team_name)
-		{
-			xform_printf(OF, fmt_team_name,
-				xform_name(player->name,server), RD,
-				player->score, RD,
-				player->ping, RD,
-				player->team_name, RD,
-				player->skin ? player->skin: "", RD,
-				play_time(player->connect_time,1)
-			);
-		}
-		else
-		{
-			xform_printf(OF, fmt,
-				xform_name(player->name, server), RD,
-				player->score, RD,
-				player->ping, RD,
-				player->team, RD,
-				player->skin ? player->skin: "", RD,
-				play_time(player->connect_time, 1)
-			);
-		}
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_doom3_player_info(struct qserver *server)
-{
-	static const char *fmt = "%s""%s%d""%s%d""%s%d""%s%u";
-	static const char *fmt_team_name = "%s""%s%d""%s%d""%s%s""%s%u";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (player->tribe_tag)
-		{
-			xform_printf(OF, fmt_team_name, xform_name(player->name, server), RD, player->score, RD, player->ping, RD, player->tribe_tag, RD, player->number);
-		}
-		else
-		{
-			xform_printf(OF, fmt, xform_name(player->name, server), RD, player->score, RD, player->ping, RD, player->team, RD, player->number);
-		}
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_gs2_player_info(struct qserver *server)
-{
-	static const char *fmt = "%s""%s%d""%s%d""%s%d""%s%s""%s%s";
-	static const char *fmt_team_name = "%s""%s%d""%s%d""%s%s""%s%s""%s%s";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (player->team_name)
-		{
-			xform_printf(OF, fmt_team_name, xform_name(player->name, server), RD,
-				player->score, RD,
-				player->ping, RD,
-				player->team_name, RD,
-				player->skin ? player->skin: "", RD,
-				play_time(player->connect_time,1)
-			);
-		}
-		else
-		{
-			xform_printf(OF, fmt,
-				xform_name(player->name,server), RD,
-				player->score, RD,
-				player->ping, RD,
-				player->team, RD,
-				player->skin ? player->skin: "", RD,
-				play_time(player->connect_time,1)
-			);
-		}
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_armyops_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		player->score = calculate_armyops_score(player);
-	}
-
-	raw_display_gs2_player_info(server);
-}
-
-void raw_display_ts2_player_info(struct qserver *server)
-{
-	static const char *fmt = "%s""%s%d""%s%s""%s%s";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt,
-			xform_name(player->name,server), RD,
-			player->ping, RD,
-			player->skin ? player->skin: "", RD,
-			play_time(player->connect_time, 1)
-		);
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_ts3_player_info(struct qserver *server)
-{
-	static const char *fmt = "%s""%s%s""%s%s";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt,
-			xform_name(player->name,server), RD,
-			player->skin ? player->skin: "", RD,
-			play_time(player->connect_time, 1)
-		);
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_starmade_player_info(struct qserver *server)
-{
-	static const char *fmt = "%s""%s%s""%s%s";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt,
-			xform_name(player->name,server), RD,
-			player->skin ? player->skin: "", RD,
-			play_time(player->connect_time, 1)
-		);
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_bfbc2_player_info(struct qserver *server)
-{
-	static const char *fmt = "%s""%s%s""%s%s";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt,
-			xform_name(player->name,server), RD,
-			player->skin ? player->skin: "", RD,
-			play_time(player->connect_time, 1)
-		);
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_wic_player_info(struct qserver *server)
-{
-	static const char *fmt = "%s""%s%d""%s%s""%s%s";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt,
-			xform_name(player->name,server), RD,
-			player->score, RD,
-			player->team_name, RD,
-			player->tribe_tag ? player->tribe_tag : ""
-		);
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_ventrilo_player_info(struct qserver *server)
-{
-	static const char *fmt = "%s""%s%d""%s%s""%s%s";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		fprintf(
-			OF, fmt,
-			xform_name(player->name, server),
-			RD, player->team,
-			RD,	player->team_name,
-			RD,	play_time(player->connect_time, 1)
-		);
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_tm_player_info(struct qserver *server)
-{
-	static const char *fmt = "%s""%s%d""%s%s""%s%s";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt,
-			xform_name(player->name,server), RD,
-			player->ping, RD,
-			player->skin ? player->skin: "", RD,
-			play_time(player->connect_time,1)
-		);
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_tee_player_info(struct qserver *server)
-{
-	static const char *fmt = "%s";
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt, xform_name(player->name,server) );
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_ravenshield_player_info(struct qserver *server)
-{
-	static char fmt[24] = "%s""%s%d""%s%s";
-	struct player *player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt, xform_name(player->name, server), RD, player->frags, RD, play_time(player->connect_time, 1));
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_savage_player_info(struct qserver *server)
-{
-	static char fmt[24] = "%s""%s%d""%s%s";
-	struct player *player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt, xform_name(player->name, server), RD, player->frags, RD, play_time(player->connect_time, 1));
-		fputs("\n", OF);
-	}
-}
-
-void raw_display_farcry_player_info(struct qserver *server)
-{
-	static char fmt[24] = "%s""%s%d""%s%s";
-	struct player *player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, fmt, xform_name(player->name, server), RD, player->frags, RD, play_time(player->connect_time, 1));
-		fputs("\n", OF);
-	}
-}
-
-/* XML output
- * Contributed by <sgarner@gameplanet.co.nz> :-)
- */
-
-void xml_display_server(struct qserver *server)
-{
-	char *prefix;
-	prefix = server->type->type_prefix;
-
-	if (server->server_name == DOWN)
-	{
-		if (!up_servers_only)
-		{
-			xform_printf(OF, "\t<server type=\"%s\" address=\"%s\" status=\"%s\">\n", xml_escape(prefix), xml_escape(server->arg), xml_escape(DOWN));
-			xform_printf(OF, "\t\t<hostname>%s</hostname>\n", xml_escape((hostname_lookup) ? server->host_name: server->arg));
-			xform_printf(OF, "\t</server>\n");
-		}
-		return;
-	}
-	if (server->server_name == TIMEOUT)
-	{
-		if (server->flags &FLAG_BROADCAST && server->n_servers)
-		{
-			xform_printf(OF, "\t<server type=\"%s\" address=\"%s\" status=\"%s\" servers=\"%d\">\n",
-				xml_escape(prefix),
-				xml_escape(server->arg),
-				xml_escape(TIMEOUT),
-				server->n_servers
-			);
-			xform_printf(OF, "\t</server>\n");
-		}
-		else if (!up_servers_only)
-		{
-			xform_printf(OF, "\t<server type=\"%s\" address=\"%s\" status=\"%s\">\n", xml_escape(prefix), xml_escape(server->arg), xml_escape(TIMEOUT));
-			xform_printf(OF, "\t\t<hostname>%s</hostname>\n", xml_escape((hostname_lookup) ? server->host_name: server->arg));
-			xform_printf(OF, "\t</server>\n");
-		}
-		return ;
-	}
-
-	if (server->error != NULL)
-	{
-		xform_printf(OF, "\t<server type=\"%s\" address=\"%s\" status=\"%s\">\n", xml_escape(prefix), xml_escape(server->arg), "ERROR");
-		xform_printf(OF, "\t\t<hostname>%s</hostname>\n", xml_escape((hostname_lookup) ? server->host_name: server->arg));
-		xform_printf(OF, "\t\t<error>%s</error>\n", xml_escape(server->error));
-	}
-	else if (server->type->master)
-	{
-		xform_printf(OF, "\t<server type=\"%s\" address=\"%s\" status=\"%s\" servers=\"%d\">\n", xml_escape(prefix), xml_escape(server->arg), "UP", server->n_servers);
-	}
-	else
-	{
-		xform_printf(OF, "\t<server type=\"%s\" address=\"%s\" status=\"%s\">\n", xml_escape(prefix), xml_escape(server->arg), "UP");
-		xform_printf(OF, "\t\t<hostname>%s</hostname>\n", xml_escape((hostname_lookup) ? server->host_name: server->arg));
-		xform_printf(OF, "\t\t<name>%s</name>\n", xml_escape(xform_name(server->server_name, server)));
-		xform_printf(OF, "\t\t<gametype>%s</gametype>\n", xml_escape(get_qw_game(server)));
-		xform_printf(OF, "\t\t<map>%s</map>\n", xml_escape(xform_name(server->map_name, server)));
-		xform_printf(OF, "\t\t<numplayers>%d</numplayers>\n", server->num_players);
-		xform_printf(OF, "\t\t<maxplayers>%d</maxplayers>\n", server->max_players);
-		xform_printf(OF, "\t\t<numspectators>%d</numspectators>\n", server->num_spectators);
-		xform_printf(OF, "\t\t<maxspectators>%d</maxspectators>\n", server->max_spectators);
-
-		if (!(server->type->flags &TF_RAW_STYLE_TRIBES))
-		{
-			xform_printf(OF, "\t\t<ping>%d</ping>\n", server->n_requests ? server->ping_total / server->n_requests: 999);
-			xform_printf(OF, "\t\t<retries>%d</retries>\n", server->n_retries);
-		}
-
-		if (server->type->flags &TF_RAW_STYLE_QUAKE)
-		{
-			xform_printf(OF, "\t\t<address>%s</address>\n", xml_escape(server->address));
-			xform_printf(OF, "\t\t<protocolversion>%d</protocolversion>\n", server->protocol_version);
-		}
-	}
-
-	if (!server->type->master && server->error == NULL)
-	{
-		if (get_server_rules && NULL != server->type->display_xml_rule_func )
-		{
-			server->type->display_xml_rule_func(server);
-		}
-		if (get_player_info && NULL != server->type->display_xml_player_func )
-		{
-			server->type->display_xml_player_func(server);
-		}
-	}
-
-	xform_printf(OF, "\t</server>\n");
-}
-
-void xml_header()
-{
-	if (xml_encoding == ENCODING_LATIN_1)
-	{
-		xform_printf(OF, "<?xml version=\"1.0\" encoding=\"iso-8859-1\"?>\n<qstat>\n");
-	}
-	else if (output_bom)
-	{
-		xform_printf(OF, "%c%c%c<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<qstat>\n", 0xEF, 0xBB, 0xBF);
-	}
-	else
-	{
-		xform_printf(OF, "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<qstat>\n");
-	}
-}
-
-void xml_footer()
-{
-	xform_printf(OF, "</qstat>\n");
-}
-
-void xml_display_server_rules(struct qserver *server)
-{
-	struct rule *rule;
-	rule = server->rules;
-
-	xform_printf(OF, "\t\t<rules>\n");
-	for (; rule != NULL; rule = rule->next)
-	{
-		xform_printf(OF, "\t\t\t<rule name=\"%s\">%s</rule>\n", xml_escape(rule->name), xml_escape(rule->value));
-	}
-	xform_printf(OF, "\t\t</rules>\n");
-}
-
-void xml_display_q_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player number=\"%d\">\n", player->number);
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<address>%s</address>\n", xml_escape(player->address));
-		xform_printf(OF, "\t\t\t\t<score>%d</score>\n", player->frags);
-		xform_printf(OF, "\t\t\t\t<time>%s</time>\n", xml_escape(play_time(player->connect_time, 2)));
-
-		if (color_names)
-		{
-			xform_printf(OF, "\t\t\t\t<color for=\"shirt\">%s</color>\n", xml_escape(quake_color(player->shirt_color)));
-			xform_printf(OF, "\t\t\t\t<color for=\"pants\">%s</color>\n", xml_escape(quake_color(player->pants_color)));
-		}
-		else
-		{
-			xform_printf(OF, "\t\t\t\t<color for=\"shirt\">%s</color>\n", quake_color(player->shirt_color));
-			xform_printf(OF, "\t\t\t\t<color for=\"pants\">%s</color>\n", quake_color(player->pants_color));
-		}
-
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_qw_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player number=\"%d\">\n", player->number);
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<score>%d</score>\n", player->frags);
-		xform_printf(OF, "\t\t\t\t<time>%s</time>\n", xml_escape(play_time(player->connect_time, 2)));
-
-		if (color_names)
-		{
-			xform_printf(OF, "\t\t\t\t<color for=\"shirt\">%s</color>\n", xml_escape(quake_color(player->shirt_color)));
-			xform_printf(OF, "\t\t\t\t<color for=\"pants\">%s</color>\n", xml_escape(quake_color(player->pants_color)));
-		}
-		else
-		{
-			xform_printf(OF, "\t\t\t\t<color for=\"shirt\">%s</color>\n", quake_color(player->shirt_color));
-			xform_printf(OF, "\t\t\t\t<color for=\"pants\">%s</color>\n", quake_color(player->pants_color));
-		}
-
-		xform_printf(OF, "\t\t\t\t<ping>%d</ping>\n", player->ping);
-		xform_printf(OF, "\t\t\t\t<skin>%s</skin>\n", player->skin ? xml_escape(player->skin): "");
-		xform_printf(OF, "\t\t\t\t<team>%s</team>\n", player->team_name ? xml_escape(player->team_name): "");
-
-
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_q2_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<score>%d</score>\n", player->frags);
-		if (server->flags &FLAG_PLAYER_TEAMS)
-		{
-			xform_printf(OF, "\t\t\t\t<team>%d</team>\n", player->team);
-		}
-		xform_printf(OF, "\t\t\t\t<ping>%d</ping>\n", player->ping);
-
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_player_info_info(struct player *player)
-{
-	struct info *info;
-
-	for (info = player->info; info; info = info->next)
-	{
-		if (info->name)
-		{
-			char *name = xml_escape(info->name);
-			char *value = xml_escape(info->value);
-			xform_printf(OF, "\t\t\t\t<%s>%s</%s>\n", name, value, name);
-		}
-	}
-}
-
-void xml_display_unreal_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<score>%d</score>\n", player->frags);
-		if ( - 999 != player->deaths)
-		{
-			xform_printf(OF, "\t\t\t\t<deaths>%d</deaths>\n", player->deaths);
-		} xform_printf(OF, "\t\t\t\t<ping>%d</ping>\n", player->ping);
-
-		if (player->team_name)
-		{
-			xform_printf(OF, "\t\t\t\t<team>%s</team>\n", xml_escape(player->team_name));
-		}
-		else if ( - 1 != player->team)
-		{
-			xform_printf(OF, "\t\t\t\t<team>%d</team>\n", player->team);
-		}
-
-		// Some games dont provide
-		// so only display if they do
-		if (player->skin)
-		{
-			xform_printf(OF, "\t\t\t\t<skin>%s</skin>\n", player->skin ? xml_escape(player->skin): "");
-		}
-		if (player->mesh)
-		{
-			xform_printf(OF, "\t\t\t\t<mesh>%s</mesh>\n", player->mesh ? xml_escape(player->mesh): "");
-		}
-		if (player->face)
-		{
-			xform_printf(OF, "\t\t\t\t<face>%s</face>\n", player->face ? xml_escape(player->face): "");
-		}
-
-		xml_display_player_info_info(player);
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_halflife_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<score>%d</score>\n", player->frags);
-		xform_printf(OF, "\t\t\t\t<time>%s</time>\n", xml_escape(play_time(player->connect_time, 2)));
-
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_fl_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<score>%d</score>\n", player->frags);
-		xform_printf(OF, "\t\t\t\t<ping>%d</ping>\n", player->ping);
-		xform_printf(OF, "\t\t\t\t<team>%d</team>\n", player->team);
-		xform_printf(OF, "\t\t\t\t<time>%s</time>\n", xml_escape(play_time(player->connect_time, 2)));
-
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_tribes_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<score>%d</score>\n", player->frags);
-		xform_printf(OF, "\t\t\t\t<team>%d</team>\n", player->team);
-		xform_printf(OF, "\t\t\t\t<ping>%d</ping>\n", player->ping);
-		xform_printf(OF, "\t\t\t\t<packetloss>%d</packetloss>\n", player->packet_loss);
-
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_tribes2_player_info(struct qserver *server)
-{
-	struct player *player;
-	char *type;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (player->team_name)
-		{
-			switch (player->type_flag)
-			{
-				case PLAYER_TYPE_BOT:
-					type = "Bot";
-					break;
-				case PLAYER_TYPE_ALIAS:
-					type = "Alias";
-					break;
-				default:
-					type = "";
-					break;
-			}
-
-			xform_printf(OF, "\t\t\t<player>\n");
-
-			xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-			xform_printf(OF, "\t\t\t\t<score>%d</score>\n", player->frags);
-			xform_printf(OF, "\t\t\t\t<team number=\"%d\">%s</team>\n", player->team, xml_escape(player->team_name));
-			xform_printf(OF, "\t\t\t\t<type>%s</type>\n", xml_escape(type));
-			xform_printf(OF, "\t\t\t\t<clan>%s</clan>\n", player->tribe_tag ? xml_escape(xform_name(player->tribe_tag, server)): "");
-
-			xform_printf(OF, "\t\t\t</player>\n");
-		}
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_bfris_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player number=\"%d\">\n", player->number);
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<score type=\"score\">%d</score>\n", player->score);
-		xform_printf(OF, "\t\t\t\t<score type=\"frags\">%d</score>\n", player->frags);
-		xform_printf(OF, "\t\t\t\t<team>%s</team>\n", xml_escape(player->team_name));
-		xform_printf(OF, "\t\t\t\t<ping>%d</ping>\n", player->ping);
-		xform_printf(OF, "\t\t\t\t<ship>%d</ship>\n", player->ship);
-
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_descent3_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<score>%d</score>\n", player->frags);
-		xform_printf(OF, "\t\t\t\t<deaths>%d</deaths>\n", player->deaths);
-		xform_printf(OF, "\t\t\t\t<ping>%d</ping>\n", player->ping);
-		xform_printf(OF, "\t\t\t\t<team>%d</team>\n", player->team);
-
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_ravenshield_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<score>%d</score>\n", player->frags);
-		xform_printf(OF, "\t\t\t\t<time>%s</time>\n", xml_escape(play_time(player->connect_time, 2)));
-
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-
-void xml_display_ghostrecon_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<deaths>%d</deaths>\n", player->deaths);
-		xform_printf(OF, "\t\t\t\t<team>%d</team>\n", player->team);
-
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_eye_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<score>%d</score>\n", player->score);
-		xform_printf(OF, "\t\t\t\t<ping>%d</ping>\n", player->ping);
-		if (player->team_name)
-		{
-			xform_printf(OF, "\t\t\t\t<team>%s</team>\n", xml_escape(player->team_name));
-		}
-		else
-		{
-			xform_printf(OF, "\t\t\t\t<team>%d</team>\n", player->team);
-		}
-		if (player->skin)
-		{
-			xform_printf(OF, "\t\t\t\t<skin>%s</skin>\n", xml_escape(player->skin));
-		}
-		if (player->connect_time)
-		{
-			xform_printf(OF, "\t\t\t\t<time>%s</time>\n", xml_escape(play_time(player->connect_time, 1)));
-		}
-
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_doom3_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<number>%u</number>\n", player->number);
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<score>%d</score>\n", player->score);
-		xform_printf(OF, "\t\t\t\t<ping>%d</ping>\n", player->ping);
-		if (player->tribe_tag)
-		{
-			xform_printf(OF, "\t\t\t\t<clan>%s</clan>\n", player->tribe_tag ? xml_escape(xform_name(player->tribe_tag, server)): "");
-		}
-		else
-		{
-			xform_printf(OF, "\t\t\t\t<team>%d</team>\n", player->team);
-		}
-		if (player->skin)
-		{
-			xform_printf(OF, "\t\t\t\t<skin>%s</skin>\n", xml_escape(player->skin));
-		}
-		if (player->type_flag)
-		{
-			xform_printf(OF, "\t\t\t\t<type>bot</type>\n");
-		}
-		else
-		{
-			xform_printf(OF, "\t\t\t\t<type>player</type>\n");
-		}
-
-		if (player->connect_time)
-		{
-			xform_printf(OF, "\t\t\t\t<time>%s</time>\n", xml_escape(play_time(player->connect_time, 2)));
-		}
-
-		xml_display_player_info_info(player);
-
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		if (NA_INT != player->ping)
-		{
-			xform_printf(OF, "\t\t\t\t<ping>%d</ping>\n", player->ping);
-		}
-		if (NA_INT != player->score)
-		{
-			xform_printf(OF, "\t\t\t\t<score>%d</score>\n", player->score);
-		}
-		if (NA_INT != player->deaths)
-		{
-			xform_printf(OF, "\t\t\t\t<deaths>%d</deaths>\n", player->deaths);
-		}
-		if (NA_INT != player->frags)
-		{
-			xform_printf(OF, "\t\t\t\t<frags>%d</frags>\n", player->frags);
-		}
-		if (player->team_name)
-		{
-			xform_printf(OF, "\t\t\t\t<team>%s</team>\n", xml_escape(player->team_name));
-		}
-		else if (NA_INT != player->team)
-		{
-			xform_printf(OF, "\t\t\t\t<team>%d</team>\n", player->team);
-		}
-
-		if (player->skin)
-		{
-			xform_printf(OF, "\t\t\t\t<skin>%s</skin>\n", xml_escape(player->skin));
-		}
-
-		if (player->connect_time)
-		{
-			xform_printf(OF, "\t\t\t\t<time>%s</time>\n", xml_escape(play_time(player->connect_time, 1)));
-		}
-
-		xml_display_player_info_info(player);
-
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_armyops_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		player->score = calculate_armyops_score(player);
-	}
-
-	xml_display_player_info(server);
-}
-
-void xml_display_ts2_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<ping>%d</ping>\n", player->ping);
-
-		if (player->connect_time)
-		{
-			xform_printf(OF, "\t\t\t\t<time>%s</time>\n", xml_escape(play_time(player->connect_time, 2)));
-		}
-
-		xml_display_player_info_info(player);
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_ts3_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-
-		if (player->connect_time)
-		{
-			xform_printf(OF, "\t\t\t\t<time>%s</time>\n", xml_escape(play_time(player->connect_time, 2)));
-		}
-
-		xml_display_player_info_info(player);
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_starmade_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-
-		if (player->connect_time)
-		{
-			xform_printf(OF, "\t\t\t\t<time>%s</time>\n", xml_escape(play_time(player->connect_time, 2)));
-		}
-
-		xml_display_player_info_info(player);
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_bfbc2_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-
-		if (player->connect_time)
-		{
-			xform_printf(OF, "\t\t\t\t<time>%s</time>\n", xml_escape(play_time(player->connect_time, 2)));
-		}
-
-		xml_display_player_info_info(player);
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_wic_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<score>%d</score>\n", player->score);
-		xform_printf(OF, "\t\t\t\t<team>%s</team>\n", player->team_name);
-		xform_printf(OF, "\t\t\t\t<bot>%d</bot>\n", player->type_flag );
-		if ( player->tribe_tag )
-		{
-			xform_printf(OF, "\t\t\t\t<role>%s</role>\n", player->tribe_tag );
-		}
-
-
-		xml_display_player_info_info(player);
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_ventrilo_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<ping>%d</ping>\n", player->ping);
-		xform_printf(OF, "\t\t\t\t<team>%s</team>\n", xml_escape(player->team_name));
-		xform_printf(OF, "\t\t\t\t<time>%s</time>\n", xml_escape(play_time(player->connect_time, 2)));
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_tm_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<ping>%d</ping>\n", player->ping);
-
-		if (player->connect_time)
-		{
-			xform_printf(OF, "\t\t\t\t<time>%s</time>\n", xml_escape(play_time(player->connect_time, 2)));
-		}
-
-		xml_display_player_info_info(player);
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-
-void xml_display_savage_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<score>%d</score>\n", player->frags);
-		xform_printf(OF, "\t\t\t\t<time>%s</time>\n", xml_escape(play_time(player->connect_time, 2)));
-
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_farcry_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<score>%d</score>\n", player->frags);
-		xform_printf(OF, "\t\t\t\t<time>%su</time>\n", xml_escape(play_time(player->connect_time, 2)));
-
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-void xml_display_tee_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	xform_printf(OF, "\t\t<players>\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		xform_printf(OF, "\t\t\t<player>\n");
-
-		xform_printf(OF, "\t\t\t\t<name>%s</name>\n", xml_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t<score>%d</score>\n", player->score);
-
-		xform_printf(OF, "\t\t\t</player>\n");
-	}
-
-	xform_printf(OF, "\t\t</players>\n");
-}
-
-/**
- * JSON output
- * Contributed by Steve Teuber <steve@exprojects.org>
- */
-
-void json_protocols()
-{
-	int i;
-	server_type *type;
-	server_type **sorted_types;
-
-	sorted_types = (server_type **)malloc(sizeof(server_type*) * n_server_types);
-	type = &types[0];
-	for (i = 0; type->id != Q_UNKNOWN_TYPE; type++, i++)
-	{
-		sorted_types[i] = type;
-	}
-
-	quicksort((void **)sorted_types, 0, n_server_types - 1, (int(*)(void *, void*))type_option_compare);
-
-	printf("{\n");
-	for (i = 0; i < n_server_types; i++)
-	{
-		type = sorted_types[i];
-		if (i) printf(",\n");
-		printf("\t\"%s\": \"%s\"", type->type_string, type->game_name);
-	}
-	printf("\n}\n");
-
-	exit(0);
-}
-
-void json_version()
-{
-	printf("{\n\t\"version\": \"%s\"\n}\n", VERSION);
-	exit(0);
-}
-
-void json_display_server(struct qserver *server)
-{
-	char *prefix;
-	prefix = server->type->type_string;
-
-	if (server->server_name == DOWN)
-	{
-		if (!up_servers_only)
-		{
-			xform_printf(OF, (json_printed) ? ",\n\t{\n" : "\t{\n");
-			xform_printf(OF, "\t\t\"protocol\": \"%s\",\n", json_escape(prefix));
-			xform_printf(OF, "\t\t\"address\": \"%s\",\n", json_escape(server->arg));
-			xform_printf(OF, "\t\t\"status\": \"%s\",\n", json_escape("offline"));
-			xform_printf(OF, "\t\t\"hostname\": \"%s\"\n", json_escape((hostname_lookup) ? server->host_name: server->arg));
-			xform_printf(OF, "\t}");
-			json_printed = 1;
-		}
-		return;
-	}
-	if (server->server_name == TIMEOUT)
-	{
-		if (server->flags &FLAG_BROADCAST && server->n_servers)
-		{
-			xform_printf(OF, (json_printed) ? ",\n\t{\n" : "\t{\n");
-			xform_printf(OF, "\t\t\"protocol\": \"%s\",\n", json_escape(prefix));
-			xform_printf(OF, "\t\t\"address\": \"%s\",\n", json_escape(server->arg));
-			xform_printf(OF, "\t\t\"status\": \"%s\",\n", json_escape("timeout"));
-			xform_printf(OF, "\t\t\"servers\": %d\n", server->n_servers);
-			xform_printf(OF, "\t}");
-			json_printed = 1;
-		}
-		else if (!up_servers_only)
-		{
-			xform_printf(OF, (json_printed) ? ",\n\t{\n" : "\t{\n");
-			xform_printf(OF, "\t\t\"protocol\": \"%s\",\n", json_escape(prefix));
-			xform_printf(OF, "\t\t\"address\": \"%s\",\n", json_escape(server->arg));
-			xform_printf(OF, "\t\t\"status\": \"%s\",\n", json_escape("timeout"));
-			xform_printf(OF, "\t\t\"hostname\": \"%s\"\n", json_escape((hostname_lookup) ? server->host_name: server->arg));
-			xform_printf(OF, "\t}");
-			json_printed = 1;
-		}
-		return ;
-	}
-
-	if (server->error != NULL)
-	{
-		xform_printf(OF, (json_printed) ? ",\n\t{\n" : "\t{\n");
-		xform_printf(OF, "\t\t\"protocol\": \"%s\",\n", json_escape(prefix));
-		xform_printf(OF, "\t\t\"address\": \"%s\",\n", json_escape(server->arg));
-		xform_printf(OF, "\t\t\"status\": \"%s\",\n", json_escape("error"));
-		xform_printf(OF, "\t\t\"hostname\": \"%s\"\n", json_escape((hostname_lookup) ? server->host_name: server->arg));
-		xform_printf(OF, "\t\t\"error\": \"%s\",\n", json_escape(server->error));
-		xform_printf(OF, "\t}");
-		json_printed = 1;
-	}
-	else if (server->type->master)
-	{
-		xform_printf(OF, (json_printed) ? ",\n\t{\n" : "\t{\n");
-		xform_printf(OF, "\t\t\"protocol\": \"%s\",\n", json_escape(prefix));
-		xform_printf(OF, "\t\t\"address\": \"%s\",\n", json_escape(server->arg));
-		xform_printf(OF, "\t\t\"status\": \"%s\",\n", json_escape("online"));
-		xform_printf(OF, "\t\t\"servers\": %d,\n", server->n_servers);
-		json_printed = 1;
-	}
-	else
-	{
-		xform_printf(OF, (json_printed) ? ",\n\t{\n" : "\t{\n");
-		xform_printf(OF, "\t\t\"protocol\": \"%s\",\n", json_escape(prefix));
-		xform_printf(OF, "\t\t\"address\": \"%s\",\n", json_escape(server->arg));
-		xform_printf(OF, "\t\t\"status\": \"%s\",\n", json_escape("online"));
-		xform_printf(OF, "\t\t\"hostname\": \"%s\",\n", json_escape((hostname_lookup) ? server->host_name: server->arg));
-		xform_printf(OF, "\t\t\"name\": \"%s\",\n", json_escape(xform_name(server->server_name, server)));
-		xform_printf(OF, "\t\t\"gametype\": \"%s\",\n", json_escape(get_qw_game(server)));
-		xform_printf(OF, "\t\t\"map\": \"%s\",\n", json_escape(xform_name(server->map_name, server)));
-		xform_printf(OF, "\t\t\"numplayers\": %d,\n", server->num_players);
-		xform_printf(OF, "\t\t\"maxplayers\": %d,\n", server->max_players);
-		xform_printf(OF, "\t\t\"numspectators\": %d,\n", server->num_spectators);
-		xform_printf(OF, "\t\t\"maxspectators\": %d", server->max_spectators);
-		json_printed = 1;
-
-		if (!(server->type->flags &TF_RAW_STYLE_TRIBES))
-		{
-			xform_printf(OF, ",\n\t\t\"ping\": %d,\n", server->n_requests ? server->ping_total / server->n_requests: 999);
-			xform_printf(OF, "\t\t\"retries\": %d", server->n_retries);
-		}
-
-		if (server->type->flags &TF_RAW_STYLE_QUAKE)
-		{
-			xform_printf(OF, ",\n\t\t\"address\": %s,\n", json_escape(server->address));
-			xform_printf(OF, "\t\t\"protocolversion\": %d", server->protocol_version);
-		}
-	}
-
-	if (!server->type->master && server->error == NULL)
-	{
-		if (get_server_rules && NULL != server->type->display_json_rule_func)
-		{
-			server->type->display_json_rule_func(server);
-		}
-		if (get_player_info && NULL != server->type->display_json_player_func)
-		{
-			server->type->display_json_player_func(server);
-		}
-	}
-
-	xform_printf(OF, "\n\t}");
-}
-
-void json_header()
-{
-	xform_printf(OF, "[\n");
-}
-
-void json_footer()
-{
-	xform_printf(OF, "\n]\n");
-}
-
-void json_display_server_rules(struct qserver *server)
-{
-	struct rule *rule;
-	int printed = 0;
-	rule = server->rules;
-
-	xform_printf(OF, ",\n\t\t\"rules\": {\n");
-	for (; rule != NULL; rule = rule->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t\"%s\": \"%s\"", json_escape(rule->name), json_escape(rule->value));
-		printed = 1;
-	}
-	xform_printf(OF, "\n\t\t}");
-}
-
-void json_display_q_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\"number\": %d,\n", player->number);
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"address\": \"%s\",\n", json_escape(player->address));
-		xform_printf(OF, "\t\t\t\t\"score\": %d,\n", player->frags);
-		xform_printf(OF, "\t\t\t\t\"time\": \"%s\"\n", json_escape(play_time(player->connect_time, 2)));
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_qw_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"number\": %d,\n", player->number);
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"score\": %d,\n", player->frags);
-		xform_printf(OF, "\t\t\t\t\"time\": \"%s\",\n", json_escape(play_time(player->connect_time, 2)));
-		xform_printf(OF, "\t\t\t\t\"ping\": %d,\n", player->ping);
-		xform_printf(OF, "\t\t\t\t\"skin\": \"%s\",\n", player->skin ? json_escape(player->skin): "");
-		xform_printf(OF, "\t\t\t\t\"team\": \"%s\"\n", player->team_name ? json_escape(player->team_name): "");
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_q2_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"score\": %d,\n", player->frags);
-		if (server->flags &FLAG_PLAYER_TEAMS)
-		{
-			xform_printf(OF, "\t\t\t\t\"team\": %d,\n", player->team);
-		}
-		xform_printf(OF, "\t\t\t\t\"ping\": %d\n", player->ping);
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_player_info_info(struct player *player)
-{
-	struct info *info;
-
-	for (info = player->info; info; info = info->next)
-	{
-		if (info->name)
-		{
-			char *name = json_escape(info->name);
-			char *value = json_escape(info->value);
-			xform_printf(OF, "\t\t\t\t\"%s\": \"%s\",\n", name, value);
-		}
-	}
-}
-
-void json_display_unreal_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"score\": %d,\n", player->frags);
-		if (player->deaths != -999)
-		{
-			xform_printf(OF, "\t\t\t\t\"deaths\": %d,\n", player->deaths);
-		}
-		if (player->team_name != NULL)
-		{
-			xform_printf(OF, "\t\t\t\t\"team\": \"%s\",\n", json_escape(player->team_name));
-		}
-		else if (player->team != -1)
-		{
-			xform_printf(OF, "\t\t\t\t\"team\": %d,\n", player->team);
-		}
-
-		if (player->skin != NULL)
-		{
-			xform_printf(OF, "\t\t\t\t\"skin\": \"%s\",\n", player->skin ? json_escape(player->skin): "");
-		}
-		if (player->mesh != NULL)
-		{
-			xform_printf(OF, "\t\t\t\t\"mesh\": \"%s\",\n", player->mesh ? json_escape(player->mesh): "");
-		}
-		if (player->face != NULL)
-		{
-			xform_printf(OF, "\t\t\t\t\"face\": \"%s\",\n", player->face ? json_escape(player->face): "");
-		}
-		json_display_player_info_info(player);
-		xform_printf(OF, "\t\t\t\t\"ping\": %d\n", player->ping);
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_halflife_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"score\": %d,\n", player->frags);
-		xform_printf(OF, "\t\t\t\t\"time\": \"%s\"\n", json_escape(play_time(player->connect_time, 2)));
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_fl_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"score\": %d,\n", player->frags);
-		xform_printf(OF, "\t\t\t\t\"ping\": %d,\n", player->ping);
-		xform_printf(OF, "\t\t\t\t\"team\": %d,\n", player->team);
-		xform_printf(OF, "\t\t\t\t\"time\": \"%s\"\n", json_escape(play_time(player->connect_time, 2)));
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_tribes_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"score\": %d,\n", player->frags);
-		xform_printf(OF, "\t\t\t\t\"team\": %d,\n", player->team);
-		xform_printf(OF, "\t\t\t\t\"ping\": %d,\n", player->ping);
-		xform_printf(OF, "\t\t\t\t\"packetloss\": %d\n", player->packet_loss);
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_tribes2_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-	char *type;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (player->team_name != NULL)
-		{
-			switch (player->type_flag)
-			{
-				case PLAYER_TYPE_BOT:
-					type = "Bot";
-					break;
-				case PLAYER_TYPE_ALIAS:
-					type = "Alias";
-					break;
-				default:
-					type = "";
-					break;
-			}
-
-			if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-			xform_printf(OF, "\t\t\t{\n");
-			xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-			xform_printf(OF, "\t\t\t\t\"score\": %d,\n", player->frags);
-			xform_printf(OF, "\t\t\t\t\"team\": \"%s\",\n", json_escape(player->team_name));
-			xform_printf(OF, "\t\t\t\t\"type\": \"%s\",\n", json_escape(type));
-			xform_printf(OF, "\t\t\t\t\"clan\": \"%s\"\n", player->tribe_tag ? json_escape(xform_name(player->tribe_tag, server)): "");
-			xform_printf(OF, "\t\t\t}");
-			printed = 1;
-		}
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_bfris_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"number\": %d,\n", player->number);
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"score\": %d,\n", player->score);
-		xform_printf(OF, "\t\t\t\t\"frags\": %d,\n", player->frags);
-		xform_printf(OF, "\t\t\t\t\"team\": \"%s\",\n", json_escape(player->team_name));
-		xform_printf(OF, "\t\t\t\t\"ping\": %d,\n", player->ping);
-		xform_printf(OF, "\t\t\t\t\"ship\": %d\n", player->ship);
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_descent3_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"score\": %d,\n", player->frags);
-		xform_printf(OF, "\t\t\t\t\"deaths\": %d,\n", player->deaths);
-		xform_printf(OF, "\t\t\t\t\"ping\": %d,\n", player->ping);
-		xform_printf(OF, "\t\t\t\t\"team\": %d\n", player->team);
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_ravenshield_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"score\": %d,\n", player->frags);
-		xform_printf(OF, "\t\t\t\t\"time\": \"%s\"\n", json_escape(play_time(player->connect_time, 2)));
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-
-void json_display_ghostrecon_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"deaths\": %d,\n", player->deaths);
-		xform_printf(OF, "\t\t\t\t\"team\": %d\n", player->team);
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_eye_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"score\": %d,\n", player->score);
-		if (player->team_name != NULL)
-		{
-			xform_printf(OF, "\t\t\t\t\"team\": \"%s\",\n", json_escape(player->team_name));
-		}
-		else
-		{
-			xform_printf(OF, "\t\t\t\t\"team\": %d,\n", player->team);
-		}
-		if (player->skin != NULL)
-		{
-			xform_printf(OF, "\t\t\t\t\"skin\": \"%s\",\n", json_escape(player->skin));
-		}
-		if (player->connect_time != 0)
-		{
-			xform_printf(OF, "\t\t\t\t\"time\": \"%s\",\n", json_escape(play_time(player->connect_time, 1)));
-		}
-		xform_printf(OF, "\t\t\t\t\"ping\": %d\n", player->ping);
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_doom3_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"number\": %d,\n", player->number);
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"score\": %d,\n", player->score);
-		if (player->tribe_tag != NULL)
-		{
-			xform_printf(OF, "\t\t\t\t\"clan\": \"%s\",\n", player->tribe_tag ? json_escape(xform_name(player->tribe_tag, server)): "");
-		}
-		else
-		{
-			xform_printf(OF, "\t\t\t\t\"team\": %d,\n", player->team);
-		}
-		if (player->skin != NULL)
-		{
-			xform_printf(OF, "\t\t\t\t\"skin\": \"%s\",\n", json_escape(player->skin));
-		}
-		if (player->type_flag != 0)
-		{
-			xform_printf(OF, "\t\t\t\t\"type\": \"%s\",\n", "bot");
-		}
-		else
-		{
-			xform_printf(OF, "\t\t\t\t\"type\": \"%s\",\n", "player");
-		}
-		if (player->connect_time != 0)
-		{
-			xform_printf(OF, "\t\t\t\t\"time\": \"%s\",\n", json_escape(play_time(player->connect_time, 2)));
-		}
-		json_display_player_info_info(player);
-		xform_printf(OF, "\t\t\t\t\"ping\": %d\n", player->ping);
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		if (NA_INT != player->ping)
-		{
-			xform_printf(OF, "\t\t\t\t\"ping\": %d,\n", player->ping);
-		}
-		if (NA_INT != player->score)
-		{
-			xform_printf(OF, "\t\t\t\t\"score\": %d,\n", player->score);
-		}
-		if (NA_INT != player->deaths)
-		{
-			xform_printf(OF, "\t\t\t\t\"deaths\": %d,\n", player->deaths);
-		}
-		if (NA_INT != player->frags)
-		{
-			xform_printf(OF, "\t\t\t\t\"frags\": %d,\n", player->frags);
-		}
-		if (player->team_name != NULL)
-		{
-			xform_printf(OF, "\t\t\t\t\"team\": \"%s\",\n", json_escape(player->team_name));
-		}
-		else if (NA_INT != player->team)
-		{
-			xform_printf(OF, "\t\t\t\t\"team\": %d,\n", player->team);
-		}
-		if (player->skin != NULL)
-		{
-			xform_printf(OF, "\t\t\t\t\"skin\": \"%s\",\n", json_escape(player->skin));
-		}
-		if (player->connect_time != 0)
-		{
-			xform_printf(OF, "\t\t\t\t\"time\": \"%s\",\n", json_escape(play_time(player->connect_time, 1)));
-		}
-		json_display_player_info_info(player);
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\"\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_armyops_player_info(struct qserver *server)
-{
-	struct player *player;
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		player->score = calculate_armyops_score(player);
-	}
-
-	json_display_player_info(server);
-}
-
-void json_display_ts2_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		if (player->connect_time != 0)
-		{
-			xform_printf(OF, "\t\t\t\t\"time\": \"%s\",\n", json_escape(play_time(player->connect_time, 2)));
-		}
-		json_display_player_info_info(player);
-		xform_printf(OF, "\t\t\t\t\"ping\": %d\n", player->ping);
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_ts3_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		if (player->connect_time != 0)
-		{
-			xform_printf(OF, "\t\t\t\t\"time\": \"%s\",\n", json_escape(play_time(player->connect_time, 2)));
-		}
-		json_display_player_info_info(player);
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\"\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_bfbc2_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		if (player->connect_time != 0)
-		{
-			xform_printf(OF, "\t\t\t\t\"time\": \"%s\",\n", json_escape(play_time(player->connect_time, 2)));
-		}
-		json_display_player_info_info(player);
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\"\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_wic_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"score\": %d,\n", player->score);
-		xform_printf(OF, "\t\t\t\t\"team\": \"%s\",\n", json_escape(player->team_name));
-		if (player->tribe_tag != NULL)
-		{
-			xform_printf(OF, "\t\t\t\t\"role\": \"%s\",\n", json_escape(player->tribe_tag));
-		}
-		json_display_player_info_info(player);
-		xform_printf(OF, "\t\t\t\t\"bot\": %d\n", player->type_flag);
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_ventrilo_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"ping\": %d,\n", player->ping);
-		xform_printf(OF, "\t\t\t\t\"team\": \"%s\",\n", json_escape(player->team_name));
-		xform_printf(OF, "\t\t\t\t\"time\": \"%s\"\n", json_escape(play_time(player->connect_time, 2)));
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_tm_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		if (player->connect_time != 0)
-		{
-			xform_printf(OF, "\t\t\t\t\"time\": \"%s\",\n", json_escape(play_time(player->connect_time, 2)));
-		}
-		json_display_player_info_info(player);
-		xform_printf(OF, "\t\t\t\t\"ping\": %d\n", player->ping);
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-
-void json_display_savage_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"score\": %d,\n", player->frags);
-		xform_printf(OF, "\t\t\t\t\"time\": \"%s\"\n", json_escape(play_time(player->connect_time, 2)));
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_farcry_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"score\": %d,\n", player->frags);
-		xform_printf(OF, "\t\t\t\t\"time\": \"%s\"\n", json_escape(play_time(player->connect_time, 2)));
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_tee_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		xform_printf(OF, "\t\t\t\t\"score\": %d\n", player->score);
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
-}
-
-void json_display_starmade_player_info(struct qserver *server)
-{
-	struct player *player;
-	int printed = 0;
-
-	xform_printf(OF, ",\n\t\t\"players\": [\n");
-
-	player = server->players;
-	for (; player != NULL; player = player->next)
-	{
-		if (printed)
-		{
-			xform_printf(OF, ",\n");
-		}
-		xform_printf(OF, "\t\t\t{\n");
-		xform_printf(OF, "\t\t\t\t\"name\": \"%s\",\n", json_escape(xform_name(player->name, server)));
-		if (player->connect_time != 0)
-		{
-			xform_printf(OF, "\t\t\t\t\"time\": \"%s\",\n", json_escape(play_time(player->connect_time, 2)));
-		}
-		json_display_player_info_info(player);
-		xform_printf(OF, "\t\t\t\t\"ping\": %d\n", player->ping);
-		xform_printf(OF, "\t\t\t}");
-		printed = 1;
-	}
-
-	xform_printf(OF, "\n\t\t]");
 }
 
 
@@ -3605,18 +326,63 @@ void display_progress()
 	}
 }
 
+int calculate_armyops_score(struct player *player)
+{
+	/* Calculates a player's score for ArmyOps from the basic components */
 
-/* ----- END MODIFICATION ----- Don't need to change anything below here. */
+	int score = 0;
+	int kill_score = 0;
+	struct info *info;
 
+	for (info = player->info; info; info = info->next)
+	{
+		if (0 == strcmp(info->name, "leader") || 0 == strcmp(info->name, "goal") || 0 == strcmp(info->name, "roe"))
+		{
+			score += atoi(info->value);
+		}
+		else if (0 == strcmp(info->name, "kia") || 0 == strcmp(info->name, "enemy"))
+		{
+			kill_score += atoi(info->value);
+		}
+	}
 
-void set_non_blocking(int fd);
-int set_fds(fd_set *fds);
-void get_next_timeout(struct timeval *timeout);
+	if (kill_score > 0)
+	{
+		score += kill_score;
+	}
 
-void set_file_descriptors();
-int wait_for_file_descriptors(struct timeval *timeout);
-struct qserver *get_next_ready_server();
+	return score;
+}
 
+char *get_qw_game(struct qserver *server)
+{
+	struct rule *rule;
+	char *game_rule = server->type->game_rule;
+	if (game_rule == NULL || *game_rule == '\0')
+	{
+		return "";
+	} rule = server->rules;
+	for (; rule != NULL; rule = rule->next)
+	{
+		if (strcmp(rule->name, game_rule) == 0)
+		{
+			if (server->type->id == Q3_SERVER && strcmp(rule->value, "baseq3") == 0)
+			{
+				return "";
+			}
+			return rule->value;
+		}
+	}
+	rule = server->rules;
+	for (; rule != NULL; rule = rule->next)
+	{
+		if (0 == strcmp(rule->name, "game") || 0 == strcmp(rule->name, "fs_game"))
+		{
+			return rule->value;
+		}
+	}
+	return "";
+}
 
 /* Misc flags
  */
@@ -4981,7 +1747,7 @@ int main(int argc, char *argv[])
 	}
 	else if (new_style && !raw_display && !have_server_template())
 	{
-		display_header();
+		standard_display_header();
 	}
 	else if (have_header_template())
 	{
@@ -7617,10 +4383,12 @@ void free_rule(struct rule *rule)
 	free(rule);
 }
 
-/* Functions for handling response packets
+/* 
+ * Functions for handling response packets
  */
 
-/* Packet from normal Quake server
+/* 
+ * Packet from normal Quake server
  */
 query_status_t deal_with_q_packet(struct qserver *server, char *rawpkt, int pktlen)
 {
@@ -7671,7 +4439,8 @@ query_status_t deal_with_q_packet(struct qserver *server, char *rawpkt, int pktl
 }
 
 
-/* Packet from QuakeWorld server
+/* 
+ * Packet from QuakeWorld server
  */
 query_status_t deal_with_qw_packet(struct qserver *server, char *rawpkt, int pktlen)
 {
@@ -8310,7 +5079,8 @@ int ack_descent3master_packet(struct qserver *server, char *curtok)
 	return rc;
 }
 
-/* Packet from Descent3 master server (PXO)
+/* 
+ * Packet from Descent3 master server (PXO)
  */
 query_status_t deal_with_descent3master_packet(struct qserver *server, char *rawpkt, int pktlen)
 {
@@ -8357,7 +5127,8 @@ query_status_t deal_with_descent3master_packet(struct qserver *server, char *raw
 	return INPROGRESS;
 }
 
-/* Packet from QuakeWorld master server
+/* 
+ * Packet from QuakeWorld master server
  */
 query_status_t deal_with_qwmaster_packet(struct qserver *server, char *rawpkt, int pktlen)
 {
@@ -8593,7 +5364,8 @@ int decode_stefmaster_packet(struct qserver *server, char *pkt, int pktlen)
 	return 1;
 }
 
-/* Packet from Tribes master server
+/* 
+ * Packet from Tribes master server
  */
 query_status_t deal_with_tribesmaster_packet(struct qserver *server, char *rawpkt, int pktlen)
 {
@@ -13192,396 +9964,6 @@ void put_long_little(unsigned val, char *buf)
 	buf[3] = (val >> 24) &0xFF;
 }
 
-#define MAXSTRLEN 2048
-
-char *xml_escape(char *string)
-{
-	static unsigned char _buf[4][MAXSTRLEN + 8];
-	static int _buf_index = 0;
-	unsigned char *result, *b, *end;
-	unsigned int c;
-
-	if (string == NULL)
-	{
-		return "";
-	}
-
-	result = &_buf[_buf_index][0];
-	_buf_index = (_buf_index + 1) % 4;
-
-	end = &result[MAXSTRLEN];
-
-	b = result;
-	for (; *string && b < end; string++)
-	{
-		c = *string;
-		switch (c)
-		{
-			case '&':
-				*b++ = '&';
-				*b++ = 'a';
-				*b++ = 'm';
-				*b++ = 'p';
-				*b++ = ';';
-				continue;
-			case '\'':
-				*b++ = '&';
-				*b++ = 'a';
-				*b++ = 'p';
-				*b++ = 'o';
-				*b++ = 's';
-				*b++ = ';';
-				continue;
-			case '"':
-				*b++ = '&';
-				*b++ = 'q';
-				*b++ = 'u';
-				*b++ = 'o';
-				*b++ = 't';
-				*b++ = ';';
-				continue;
-			case '<':
-				*b++ = '&';
-				*b++ = 'l';
-				*b++ = 't';
-				*b++ = ';';
-				continue;
-			case '>':
-				*b++ = '&';
-				*b++ = 'g';
-				*b++ = 't';
-				*b++ = ';';
-				continue;
-			default:
-				break;
-		}
-
-		// Validate character
-		// http://www.w3.org/TR/2000/REC-xml-20001006#charsets
-		if ( !
-			(
-				0x09 == c ||
-				0xA == c ||
-				0xD == c ||
-				( 0x20 <= c && 0xD7FF >= c ) ||
-				( 0xE000 <= c && 0xFFFD >= c ) ||
-				( 0x10000 <= c && 0x10FFFF >= c )
-			)
-		)
-		{
-			if ( show_errors )
-			{
-				fprintf(stderr, "Encoding error (%d) for U+%x, D+%d\n", 1, c, c);
-			}
-		}
-		else if (xml_encoding == ENCODING_LATIN_1)
-		{
-			if (!xform_names)
-			{
-				*b++ = c;
-			}
-			else
-			{
-				if (isprint(c))
-				{
-					*b++ = c;
-				}
-				else
-				{
-					b += sprintf( (char *)b, "&#%u;", c);
-				}
-			}
-		}
-		else if (xml_encoding == ENCODING_UTF_8)
-		{
-			unsigned char tempbuf[10] =
-			{
-				0
-			};
-			unsigned char *buf = &tempbuf[0];
-			int bytes = 0;
-			int error = 1;
-
-			// Valid character ranges
-			if (
-				0x09 == c ||
-				0xA == c ||
-				0xD == c ||
-				( 0x20 <= c && 0xD7FF >= c ) ||
-				( 0xE000 <= c && 0xFFFD >= c ) ||
-				( 0x10000 <= c && 0x10FFFF >= c )
-			)
-			{
-				error = 0;
-			}
-
-			if (c < 0x80)
-			/* 0XXX XXXX one byte */
-			{
-				buf[0] = c;
-				bytes = 1;
-			}
-			else if (c < 0x0800)
-			/* 110X XXXX two bytes */
-			{
-				buf[0] = 0xC0 | (0x03 &(c >> 6));
-				buf[1] = 0x80 | (0x3F &c);
-				bytes = 2;
-			}
-			else if (c < 0x10000)
-			/* 1110 XXXX three bytes */
-			{
-				buf[0] = 0xE0 | (c >> 12);
-				buf[1] = 0x80 | ((c >> 6) &0x3F);
-				buf[2] = 0x80 | (c &0x3F);
-
-				bytes = 3;
-				if (c == UTF8BYTESWAPNOTACHAR || c == UTF8NOTACHAR)
-				{
-					error = 3;
-				}
-
-			}
-			else if (c < 0x10FFFF)
-			/* 1111 0XXX four bytes */
-			{
-				buf[0] = 0xF0 | (c >> 18);
-				buf[1] = 0x80 | ((c >> 12) &0x3F);
-				buf[2] = 0x80 | ((c >> 6) &0x3F);
-				buf[3] = 0x80 | (c &0x3F);
-				bytes = 4;
-				if (c > UTF8MAXFROMUCS4)
-				{
-					error = 4;
-				}
-
-			}
-			else if (c < 0x4000000)
-			/* 1111 10XX five bytes */
-			{
-				buf[0] = 0xF8 | (c >> 24);
-				buf[1] = 0x80 | (c >> 18);
-				buf[2] = 0x80 | ((c >> 12) &0x3F);
-				buf[3] = 0x80 | ((c >> 6) &0x3F);
-				buf[4] = 0x80 | (c &0x3F);
-				bytes = 5;
-				error = 5;
-			}
-			else if (c < 0x80000000)
-			/* 1111 110X six bytes */
-			{
-				buf[0] = 0xFC | (c >> 30);
-				buf[1] = 0x80 | ((c >> 24) &0x3F);
-				buf[2] = 0x80 | ((c >> 18) &0x3F);
-				buf[3] = 0x80 | ((c >> 12) &0x3F);
-				buf[4] = 0x80 | ((c >> 6) &0x3F);
-				buf[5] = 0x80 | (c &0x3F);
-				bytes = 6;
-				error = 6;
-			}
-			else
-			{
-				error = 7;
-			}
-
-			if (error)
-			{
-				int i;
-				fprintf(stderr, "UTF-8 encoding error (%d) for U+%x, D+%d : ", error, c, c);
-				for (i = 0; i < bytes; i++)
-				{
-					fprintf(stderr, "0x%02x ", buf[i]);
-				}
-				fprintf(stderr, "\n");
-			}
-			else
-			{
-				int i;
-				for (i = 0; i < bytes; ++i)
-				{
-					*b++ = buf[i];
-				}
-			}
-
-		}
-	}
-	*b = '\0';
-	return (char*)result;
-}
-
-char *json_escape(char *string)
-{
-	static unsigned char _buf[4][MAXSTRLEN + 8];
-	static int _buf_index = 0;
-	unsigned char *result, *b, *end;
-	unsigned int c;
-
-	if (string == NULL)
-	{
-		return "";
-	}
-
-	result = &_buf[_buf_index][0];
-	_buf_index = (_buf_index + 1) % 4;
-
-	end = &result[MAXSTRLEN];
-
-	b = result;
-	for (; *string && b < end; string++)
-	{
-		c = *string;
-		switch (c)
-		{
-			case '"':
-				*b++ = '\\';
-				*b++ = '"';
-				continue;
-			case '\\':
-				*b++ = '\\';
-				*b++ = '\\';
-				continue;
-			default:
-				break;
-		}
-
-		// Validate character
-		// http://www.w3.org/TR/2000/REC-xml-20001006#charsets
-		if ( !
-			(
-				0x09 == c ||
-				0xA == c ||
-				0xD == c ||
-				( 0x20 <= c && 0xD7FF >= c ) ||
-				( 0xE000 <= c && 0xFFFD >= c ) ||
-				( 0x10000 <= c && 0x10FFFF >= c )
-			)
-		)
-		{
-			if ( show_errors )
-			{
-				fprintf(stderr, "Encoding error (%d) for U+%x, D+%d\n", 1, c, c);
-			}
-		}
-		// JSON is always unicode-encoded, see RFC 4627
-		// https://www.ietf.org/rfc/rfc4627.txt
-		else
-		{
-			unsigned char tempbuf[10] =
-			{
-				0
-			};
-			unsigned char *buf = &tempbuf[0];
-			int bytes = 0;
-			int error = 1;
-
-			// Valid character ranges
-			if (
-				0x09 == c ||
-				0xA == c ||
-				0xD == c ||
-				( 0x20 <= c && 0xD7FF >= c ) ||
-				( 0xE000 <= c && 0xFFFD >= c ) ||
-				( 0x10000 <= c && 0x10FFFF >= c )
-			)
-			{
-				error = 0;
-			}
-
-			if (c < 0x80)
-			/* 0XXX XXXX one byte */
-			{
-				buf[0] = c;
-				bytes = 1;
-			}
-			else if (c < 0x0800)
-			/* 110X XXXX two bytes */
-			{
-				buf[0] = 0xC0 | (0x03 &(c >> 6));
-				buf[1] = 0x80 | (0x3F &c);
-				bytes = 2;
-			}
-			else if (c < 0x10000)
-			/* 1110 XXXX three bytes */
-			{
-				buf[0] = 0xE0 | (c >> 12);
-				buf[1] = 0x80 | ((c >> 6) &0x3F);
-				buf[2] = 0x80 | (c &0x3F);
-
-				bytes = 3;
-				if (c == UTF8BYTESWAPNOTACHAR || c == UTF8NOTACHAR)
-				{
-					error = 3;
-				}
-
-			}
-			else if (c < 0x10FFFF)
-			/* 1111 0XXX four bytes */
-			{
-				buf[0] = 0xF0 | (c >> 18);
-				buf[1] = 0x80 | ((c >> 12) &0x3F);
-				buf[2] = 0x80 | ((c >> 6) &0x3F);
-				buf[3] = 0x80 | (c &0x3F);
-				bytes = 4;
-				if (c > UTF8MAXFROMUCS4)
-				{
-					error = 4;
-				}
-
-			}
-			else if (c < 0x4000000)
-			/* 1111 10XX five bytes */
-			{
-				buf[0] = 0xF8 | (c >> 24);
-				buf[1] = 0x80 | (c >> 18);
-				buf[2] = 0x80 | ((c >> 12) &0x3F);
-				buf[3] = 0x80 | ((c >> 6) &0x3F);
-				buf[4] = 0x80 | (c &0x3F);
-				bytes = 5;
-				error = 5;
-			}
-			else if (c < 0x80000000)
-			/* 1111 110X six bytes */
-			{
-				buf[0] = 0xFC | (c >> 30);
-				buf[1] = 0x80 | ((c >> 24) &0x3F);
-				buf[2] = 0x80 | ((c >> 18) &0x3F);
-				buf[3] = 0x80 | ((c >> 12) &0x3F);
-				buf[4] = 0x80 | ((c >> 6) &0x3F);
-				buf[5] = 0x80 | (c &0x3F);
-				bytes = 6;
-				error = 6;
-			}
-			else
-			{
-				error = 7;
-			}
-
-			if (error)
-			{
-				int i;
-				fprintf(stderr, "UTF-8 encoding error (%d) for U+%x, D+%d : ", error, c, c);
-				for (i = 0; i < bytes; i++)
-				{
-					fprintf(stderr, "0x%02x ", buf[i]);
-				}
-				fprintf(stderr, "\n");
-			}
-			else
-			{
-				int i;
-				for (i = 0; i < bytes; ++i)
-				{
-					*b++ = buf[i];
-				}
-			}
-
-		}
-	}
-	*b = '\0';
-	return (char*)result;
-}
-
-
 int is_default_rule(struct rule *rule)
 {
 	if (strcmp(rule->name, "sv_maxspeed") == 0)
@@ -14138,4 +10520,3 @@ int qpartition(void **array, int a, int b, int(*compare)(void *, void*))
 		}
 	}
 }
-
