+++ qstat.h
@@ -38,7 +38,7 @@
 static int gettimeofday(struct timeval *now, void *blah)
 {
     struct timeb timeb;
-    ftime( &timeb);
+    ftime(&timeb);
     now->tv_sec= timeb.time;
     now->tv_usec= (unsigned int)timeb.millitm * 1000;
     return 0;
@@ -80,9 +80,18 @@ typedef enum {
 
 #include "qserver.h"
 
-typedef void (*DisplayFunc)( struct qserver *);
-typedef query_status_t (*QueryFunc)( struct qserver *);
-typedef query_status_t (*PacketFunc)( struct qserver *, char *rawpkt, int pktlen);
+#define ENCODING_LATIN_1 1
+#define ENCODING_UTF_8  8
+#define UTF8BYTESWAPNOTACHAR	0xFFFE
+#define UTF8NOTACHAR	0xFFFF
+#define UTF8MAXFROMUCS4	0x10FFFF
+
+// Output formats
+#include "display_standard.h"
+#include "display_template.h"
+#include "display_raw.h"
+#include "display_xml.h"
+#include "display_json.h"
 
 // Packet modules
 #include "ut2004.h"
@@ -285,176 +294,121 @@ typedef query_status_t (*PacketFunc)( struct qserver *, char *rawpkt, int pktlen
 
 #define TRIBES_TEAM	-1
 
+typedef void (*DisplayFunc)(struct qserver *);
+typedef query_status_t (*QueryFunc)(struct qserver *);
+typedef query_status_t (*PacketFunc)(struct qserver *, char *rawpkt, int pktlen);
+
 struct q_packet;
 
-/*
- * Output and formatting functions
+struct qserver *servers;
+struct qserver **last_server;
+struct qserver *last_server_bind;
+struct qserver *first_server_bind;
+
+/* 
+ * Functions 
  */
 
-void display_server( struct qserver *server);
-void display_qwmaster( struct qserver *server);
-void display_server_rules( struct qserver *server);
-void display_player_info( struct qserver *server);
-void display_q_player_info( struct qserver *server);
-void display_qw_player_info( struct qserver *server);
-void display_q2_player_info( struct qserver *server);
-void display_unreal_player_info( struct qserver *server);
-void display_shogo_player_info( struct qserver *server);
-void display_halflife_player_info( struct qserver *server);
-void display_tribes_player_info( struct qserver *server);
-void display_tribes2_player_info( struct qserver *server);
-void display_bfris_player_info( struct qserver *server);
-void display_descent3_player_info( struct qserver *server);
-void display_ravenshield_player_info( struct qserver *server);
-void display_savage_player_info( struct qserver *server);
-void display_farcry_player_info( struct qserver *server);
-void display_ghostrecon_player_info( struct qserver *server);
-void display_eye_player_info( struct qserver *server);
-void display_armyops_player_info( struct qserver *server);
-void display_gs2_player_info( struct qserver *server);
-void display_doom3_player_info( struct qserver *server);
-void display_ts2_player_info( struct qserver *server);
-void display_ts3_player_info( struct qserver *server);
-void display_bfbc2_player_info( struct qserver *server);
-void display_tm_player_info( struct qserver *server);
-void display_wic_player_info( struct qserver *server);
-void display_fl_player_info( struct qserver *server);
-void display_tee_player_info( struct qserver *server);
-void display_ventrilo_player_info( struct qserver *server);
-void display_starmade_player_info( struct qserver *server);
+void quicksort(void **array, int i, int j, int(*compare)(void *, void*));
+int qpartition(void **array, int i, int j, int(*compare)(void *, void*));
 
-void raw_display_server( struct qserver *server);
-void raw_display_server_rules( struct qserver *server);
-void raw_display_player_info( struct qserver *server);
-void raw_display_q_player_info( struct qserver *server);
-void raw_display_qw_player_info( struct qserver *server);
-void raw_display_q2_player_info( struct qserver *server);
-void raw_display_unreal_player_info( struct qserver *server);
-void raw_display_halflife_player_info( struct qserver *server);
-void raw_display_tribes_player_info( struct qserver *server);
-void raw_display_tribes2_player_info( struct qserver *server);
-void raw_display_bfris_player_info( struct qserver *server);
-void raw_display_ravenshield_player_info( struct qserver *server);
-void raw_display_savage_player_info( struct qserver *server);
-void raw_display_farcry_player_info( struct qserver *server);
-void raw_display_descent3_player_info( struct qserver *server);
-void raw_display_ghostrecon_player_info( struct qserver *server);
-void raw_display_eye_player_info( struct qserver *server);
-void raw_display_armyops_player_info( struct qserver *server);
-void raw_display_gs2_player_info( struct qserver *server);
-void raw_display_doom3_player_info( struct qserver *server);
-void raw_display_ts2_player_info( struct qserver *server);
-void raw_display_ts3_player_info( struct qserver *server);
-void raw_display_bfbc2_player_info( struct qserver *server);
-void raw_display_tm_player_info( struct qserver *server);
-void raw_display_wic_player_info( struct qserver *server);
-void raw_display_fl_player_info( struct qserver *server);
-void raw_display_tee_player_info( struct qserver *server);
-void raw_display_ventrilo_player_info( struct qserver *server);
-void raw_display_starmade_player_info( struct qserver *server);
+void sort_servers(struct qserver **array, int size);
+void sort_players(struct qserver *server);
+int server_compare(struct qserver *one, struct qserver *two);
+int player_compare(struct player *one, struct player *two);
+int type_option_compare(server_type *one, server_type *two);
+int type_string_compare(server_type *one, server_type *two);
 
-void xml_display_server( struct qserver *server);
-void xml_header();
-void xml_footer();
-void xml_display_server_rules( struct qserver *server);
-void xml_display_player_info( struct qserver *server);
-void xml_display_q_player_info( struct qserver *server);
-void xml_display_qw_player_info( struct qserver *server);
-void xml_display_q2_player_info( struct qserver *server);
-void xml_display_unreal_player_info( struct qserver *server);
-void xml_display_halflife_player_info( struct qserver *server);
-void xml_display_tribes_player_info( struct qserver *server);
-void xml_display_tribes2_player_info( struct qserver *server);
-void xml_display_ravenshield_player_info( struct qserver *server);
-void xml_display_savage_player_info( struct qserver *server);
-void xml_display_farcry_player_info( struct qserver *server);
-void xml_display_bfris_player_info( struct qserver *server);
-void xml_display_descent3_player_info( struct qserver *server);
-void xml_display_ghostrecon_player_info( struct qserver *server);
-void xml_display_eye_player_info( struct qserver *server);
-void xml_display_armyops_player_info( struct qserver *server);
-void xml_display_player_info( struct qserver *server);
-void xml_display_doom3_player_info( struct qserver *server);
-void xml_display_ts2_player_info( struct qserver *server);
-void xml_display_ts3_player_info( struct qserver *server);
-void xml_display_bfbc2_player_info( struct qserver *server);
-void xml_display_tm_player_info( struct qserver *server);
-void xml_display_wic_player_info( struct qserver *server);
-void xml_display_fl_player_info( struct qserver *server);
-void xml_display_tee_player_info( struct qserver *server);
-void xml_display_ventrilo_player_info( struct qserver *server);
-void xml_display_starmade_player_info( struct qserver *server);
-char *xml_escape( char*);
+int process_func_ret( struct qserver *server, int ret );
+int connection_inprogress();
+void clear_socketerror();
 
-void json_display_server( struct qserver *server);
-void json_header();
-void json_footer();
-void json_display_server_rules( struct qserver *server);
-void json_display_player_info( struct qserver *server);
-void json_display_q_player_info( struct qserver *server);
-void json_display_qw_player_info( struct qserver *server);
-void json_display_q2_player_info( struct qserver *server);
-void json_display_unreal_player_info( struct qserver *server);
-void json_display_halflife_player_info( struct qserver *server);
-void json_display_tribes_player_info( struct qserver *server);
-void json_display_tribes2_player_info( struct qserver *server);
-void json_display_ravenshield_player_info( struct qserver *server);
-void json_display_savage_player_info( struct qserver *server);
-void json_display_farcry_player_info( struct qserver *server);
-void json_display_bfris_player_info( struct qserver *server);
-void json_display_descent3_player_info( struct qserver *server);
-void json_display_ghostrecon_player_info( struct qserver *server);
-void json_display_eye_player_info( struct qserver *server);
-void json_display_armyops_player_info( struct qserver *server);
-void json_display_player_info( struct qserver *server);
-void json_display_doom3_player_info( struct qserver *server);
-void json_display_ts2_player_info( struct qserver *server);
-void json_display_ts3_player_info( struct qserver *server);
-void json_display_bfbc2_player_info( struct qserver *server);
-void json_display_tm_player_info( struct qserver *server);
-void json_display_wic_player_info( struct qserver *server);
-void json_display_fl_player_info( struct qserver *server);
-void json_display_tee_player_info( struct qserver *server);
-void json_display_ventrilo_player_info( struct qserver *server);
-void json_display_starmade_player_info( struct qserver *server);
-char *json_escape( char*);
+void remove_server_from_hash(struct qserver *server);
+void free_server(struct qserver *server);
+void free_player(struct player *player);
+void free_rule(struct rule *rule);
 
-query_status_t send_qserver_request_packet( struct qserver *server);
-query_status_t send_qwserver_request_packet( struct qserver *server);
-query_status_t send_ut2003_request_packet( struct qserver *server);
-query_status_t send_tribes_request_packet( struct qserver *server);
-query_status_t send_qwmaster_request_packet( struct qserver *server);
-query_status_t send_bfris_request_packet( struct qserver *server);
-query_status_t send_player_request_packet( struct qserver *server);
-query_status_t send_rule_request_packet( struct qserver *server);
-query_status_t send_savage_request_packet( struct qserver *server);
-query_status_t send_farcry_request_packet( struct qserver *server);
-query_status_t send_gamespy_master_request( struct qserver *server);
-query_status_t send_tribes2_request_packet( struct qserver *server);
-query_status_t send_tribes2master_request_packet( struct qserver *server);
-query_status_t send_hl2_request_packet( struct qserver *server);
+void display_server(struct qserver *server);
+void display_progress();
+int calculate_armyops_score(struct player *player);
 
-query_status_t deal_with_q_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_qw_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_q1qw_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_q2_packet( struct qserver *server, char *pkt, int pktlen );
-query_status_t deal_with_qwmaster_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_halflife_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_ut2003_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_tribes_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_tribesmaster_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_bfris_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_gamespy_master_response( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_ravenshield_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_savage_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_farcry_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_tribes2_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_tribes2master_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_descent3_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_descent3master_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_ghostrecon_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_eye_packet( struct qserver *server, char *pkt, int pktlen);
-query_status_t deal_with_hl2_packet( struct qserver *server, char *pkt, int pktlen);
+/*
+ * Values set by command-line arguments
+ */
+
+int hostname_lookup;
+int new_style; /* unset if -old was specified */
+int up_servers_only;
+int no_full_servers;
+int no_empty_servers;
+int no_header_display;
+int player_address;
+int max_simultaneous;
+int sendinterval;
+
+int progress;
+int num_servers_total;
+int num_players_total;
+int max_players_total;
+int num_servers_returned;
+int num_servers_timed_out;
+int num_servers_down;
+
+int player_sort;
+int server_sort;
+
+extern int xform_names;
+extern int xform_strip_unprintable;
+extern int xform_hex_player_names;
+extern int xform_hex_server_names;
+extern int xform_strip_carets;
+extern int xform_html_names;
+extern int html_mode;
+
+FILE *OF; /* output file */
+
+int display_prefix;
+
+/* 
+ * Functions for handling response packets
+ */
+
+query_status_t send_qserver_request_packet(struct qserver *server);
+query_status_t send_qwserver_request_packet(struct qserver *server);
+query_status_t send_ut2003_request_packet(struct qserver *server);
+query_status_t send_tribes_request_packet(struct qserver *server);
+query_status_t send_qwmaster_request_packet(struct qserver *server);
+query_status_t send_bfris_request_packet(struct qserver *server);
+query_status_t send_player_request_packet(struct qserver *server);
+query_status_t send_rule_request_packet(struct qserver *server);
+query_status_t send_savage_request_packet(struct qserver *server);
+query_status_t send_farcry_request_packet(struct qserver *server);
+query_status_t send_gamespy_master_request(struct qserver *server);
+query_status_t send_tribes2_request_packet(struct qserver *server);
+query_status_t send_tribes2master_request_packet(struct qserver *server);
+query_status_t send_hl2_request_packet(struct qserver *server);
+
+query_status_t deal_with_q_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_qw_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_q1qw_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_q2_packet(struct qserver *server, char *pkt, int pktlen );
+query_status_t deal_with_qwmaster_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_halflife_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_ut2003_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_tribes_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_tribesmaster_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_bfris_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_gamespy_master_response(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_ravenshield_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_savage_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_farcry_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_tribes2_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_tribes2master_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_descent3_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_descent3master_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_ghostrecon_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_eye_packet(struct qserver *server, char *pkt, int pktlen);
+query_status_t deal_with_hl2_packet(struct qserver *server, char *pkt, int pktlen);
 
 struct _server_type  {
     int id;
@@ -496,7 +450,7 @@ extern server_type builtin_types[];
 extern server_type *types;
 extern int n_server_types;
 extern server_type* default_server_type;
-server_type* find_server_type_string( char* type_string);
+server_type* find_server_type_string(char* type_string);
 
 
 #ifdef QUERY_PACKETS
@@ -836,10 +790,10 @@ unsigned char haze_player_query[] = {
 
 // Steam
 // Format:
-// 1. Request type ( 1 byte )
-// 2. Region ( 1 byte )
-// 3. ip ( string + null )
-// 4. Filter ( optional + null )
+// 1. Request type (1 byte )
+// 2. Region (1 byte )
+// 3. ip (string + null )
+// 4. Filter (optional + null )
 //
 // Regions:
 // 0 = US East Coast
@@ -942,13 +896,13 @@ server_type builtin_types[] = {
     (char*) &q_player,		/* player_packet */
     Q_HEADER_LEN+1,		/* player_len */
     (char*) &q_rule,		/* rule_packet */
-    sizeof( q_rule),		/* rule_len */
+    sizeof(q_rule),		/* rule_len */
     NULL,			/* master_packet */
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_q_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_q_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_q_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_q_player_info,	/* display_xml_player_func */
@@ -978,13 +932,13 @@ server_type builtin_types[] = {
     (char*) &q_player,		/* player_packet */
     Q_HEADER_LEN+1,		/* player_len */
     (char*) &q_rule,		/* rule_packet */
-    sizeof( q_rule),		/* rule_len */
+    sizeof(q_rule),		/* rule_len */
     NULL,			/* master_packet */
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_q_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_q_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_q_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_q_player_info,	/* display_xml_player_func */
@@ -1010,7 +964,7 @@ server_type builtin_types[] = {
     "*gamedir",			/* game_rule */
     "QUAKEWORLD",		/* template_var */
     (char*) &qw_serverstatus,	/* status_packet */
-    sizeof( qw_serverstatus),	/* status_len */
+    sizeof(qw_serverstatus),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -1019,8 +973,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_qw_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_qw_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_qw_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_qw_player_info,	/* display_xml_player_func */
@@ -1046,7 +1000,7 @@ server_type builtin_types[] = {
     "*gamedir",			/* game_rule */
     "HEXENWORLD",		/* template_var */
     (char*) &hw_serverstatus,	/* status_packet */
-    sizeof( hw_serverstatus),	/* status_len */
+    sizeof(hw_serverstatus),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -1055,8 +1009,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_qw_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_qw_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_qw_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_qw_player_info,	/* display_xml_player_func */
@@ -1082,7 +1036,7 @@ server_type builtin_types[] = {
     "gamedir",			/* game_rule */
     "QUAKE2",			/* template_var */
     (char*) &qw_serverstatus,	/* status_packet */
-    sizeof( qw_serverstatus),	/* status_len */
+    sizeof(qw_serverstatus),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -1091,8 +1045,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_q2_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_q2_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_q2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_q2_player_info,	/* display_xml_player_func */
@@ -1118,17 +1072,17 @@ server_type builtin_types[] = {
     "game",			/* game_rule */
     "QUAKE3",			/* template_var */
     (char*) &q3_serverinfo,	/* status_packet */
-    sizeof( q3_serverinfo),	/* status_len */
+    sizeof(q3_serverinfo),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     (char*) &q3_serverstatus,	/* rule_packet */
-    sizeof( q3_serverstatus),	/* rule_len */
+    sizeof(q3_serverstatus),	/* rule_len */
     NULL,			/* master_packet */
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_q2_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_q2_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_q2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_q2_player_info,	/* display_xml_player_func */
@@ -1154,7 +1108,7 @@ server_type builtin_types[] = {
     "fs_game",						/* game_rule */
     "DOOM3",						/* template_var */
     (char*) &doom3_serverinfo,		/* status_packet */
-    sizeof( doom3_serverinfo),		/* status_len */
+    sizeof(doom3_serverinfo),		/* status_len */
     NULL,							/* player_packet */
     0,								/* player_len */
     NULL,							/* rule_packet */
@@ -1163,8 +1117,8 @@ server_type builtin_types[] = {
     0,								/* master_len */
     NULL,							/* master_protocol */
     NULL,							/* master_query */
-    display_doom3_player_info,		/* display_player_func */
-    display_server_rules,			/* display_rule_func */
+    standard_display_doom3_player_info,	/* display_player_func */
+    standard_display_server_rules,		/* display_rule_func */
     raw_display_doom3_player_info,	/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_doom3_player_info,	/* display_xml_player_func */
@@ -1190,17 +1144,17 @@ server_type builtin_types[] = {
     "",								/* game_rule */
     "HL2",							/* template_var */
     (char*) &hl2_serverinfo,		/* status_packet */
-    sizeof( hl2_serverinfo),		/* status_len */
+    sizeof(hl2_serverinfo),		/* status_len */
     (char*) &hl2_playerinfo,		/* player_packet */
-    sizeof( hl2_playerinfo),		/* player_len */
+    sizeof(hl2_playerinfo),		/* player_len */
     (char*) &hl2_ruleinfo,			/* rule_packet */
-    sizeof( hl2_ruleinfo),			/* rule_len */
+    sizeof(hl2_ruleinfo),			/* rule_len */
     NULL,							/* master_packet */
     0,								/* master_len */
     NULL,							/* master_protocol */
     NULL,							/* master_query */
-    display_halflife_player_info,		/* display_player_func */
-    display_server_rules,			/* display_rule_func */
+    standard_display_halflife_player_info,		/* display_player_func */
+    standard_display_server_rules,			/* display_rule_func */
     raw_display_halflife_player_info,	/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_halflife_player_info,	/* display_xml_player_func */
@@ -1226,17 +1180,17 @@ server_type builtin_types[] = {
     "game",			/* game_rule */
     "RTCW",			/* template_var */
     (char*) &q3_serverinfo,	/* status_packet */
-    sizeof( q3_serverinfo),	/* status_len */
+    sizeof(q3_serverinfo),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     (char*) &q3_serverstatus,	/* rule_packet */
-    sizeof( q3_serverstatus),	/* rule_len */
+    sizeof(q3_serverstatus),	/* rule_len */
     NULL,			/* master_packet */
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_q2_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_q2_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_q2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_q2_player_info,	/* display_xml_player_func */
@@ -1262,17 +1216,17 @@ server_type builtin_types[] = {
     "game",			/* game_rule */
     "ELITEFORCE",		/* template_var */
     (char*) &q3_serverinfo,	/* status_packet */
-    sizeof( q3_serverinfo),	/* status_len */
+    sizeof(q3_serverinfo),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     (char*) &q3_serverstatus,	/* rule_packet */
-    sizeof( q3_serverstatus),	/* rule_len */
+    sizeof(q3_serverstatus),	/* rule_len */
     NULL,			/* master_packet */
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_q2_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_q2_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_q2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_q2_player_info,	/* display_xml_player_func */
@@ -1298,17 +1252,17 @@ server_type builtin_types[] = {
     "game",			/* game_rule */
     "JEDIKNIGHT3",		/* template_var */
     (char*) &q3_serverinfo,	/* status_packet */
-    sizeof( q3_serverinfo),	/* status_len */
+    sizeof(q3_serverinfo),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     (char*) &q3_serverstatus,	/* rule_packet */
-    sizeof( q3_serverstatus),	/* rule_len */
+    sizeof(q3_serverstatus),	/* rule_len */
     NULL,			/* master_packet */
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_q2_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_q2_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_q2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_q2_player_info,	/* display_xml_player_func */
@@ -1334,17 +1288,17 @@ server_type builtin_types[] = {
     "gametype",			/* game_rule */
     "UNREALTOURNAMENT2003",	/* template_var */
     (char*) &ut2003_basicstatus,/* status_packet */
-    sizeof( ut2003_basicstatus),/* status_len */
+    sizeof(ut2003_basicstatus),/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     (char*) &ut2003_allinfo,	/* rule_packet */
-    sizeof( ut2003_allinfo),	/* rule_len */
+    sizeof(ut2003_allinfo),	/* rule_len */
     NULL,			/* master_packet */
     0,				/* master_len */
     NULL,			/* master_protocol */
     0,				/* master_query */
-    display_unreal_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_unreal_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_unreal_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_unreal_player_info,	/* display_xml_player_func */
@@ -1370,7 +1324,7 @@ server_type builtin_types[] = {
     "gametype",			/* game_rule */
     "UNREAL",			/* template_var */
     (char*) &unreal_serverstatus,	/* status_packet */
-    sizeof( unreal_serverstatus),	/* status_len */
+    sizeof(unreal_serverstatus),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -1379,8 +1333,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_unreal_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_unreal_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_unreal_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_unreal_player_info,	/* display_xml_player_func */
@@ -1406,17 +1360,17 @@ server_type builtin_types[] = {
     "game",			/* game_rule */
     "HALFLIFE",			/* template_var */
     (char*) &hl_details,		/* status_packet */
-    sizeof( hl_details),		/* status_len */
+    sizeof(hl_details),		/* status_len */
     (char*) &hl_players,	/* player_packet */
-    sizeof( hl_players),	/* player_len */
+    sizeof(hl_players),	/* player_len */
     (char*) &hl_rules,		/* rule_packet */
-    sizeof( hl_rules),		/* rule_len */
+    sizeof(hl_rules),		/* rule_len */
     NULL,			/* master_packet */
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_halflife_player_info,/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_halflife_player_info,/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_halflife_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_halflife_player_info,	/* display_xml_player_func */
@@ -1442,7 +1396,7 @@ server_type builtin_types[] = {
     "gamedir",			/* game_rule */
     "SIN",			/* template_var */
     (char*) &qw_serverstatus,	/* status_packet */
-    sizeof( qw_serverstatus),	/* status_len */
+    sizeof(qw_serverstatus),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -1451,8 +1405,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_q2_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_q2_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_q2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_q2_player_info,	/* display_xml_player_func */
@@ -1478,7 +1432,7 @@ server_type builtin_types[] = {
     "",				/* game_rule */
     "SHOGO",			/* template_var */
     (char*) &unreal_serverstatus,	/* status_packet */
-    sizeof( unreal_serverstatus),	/* status_len */
+    sizeof(unreal_serverstatus),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -1487,8 +1441,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_q2_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_q2_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_q2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_q2_player_info,	/* display_xml_player_func */
@@ -1514,17 +1468,17 @@ server_type builtin_types[] = {
     "game",			/* game_rule */
     "TRIBES",			/* template_var */
     (char*) &tribes_info,	/* status_packet */
-    sizeof( tribes_info),	/* status_len */
+    sizeof(tribes_info),	/* status_len */
     (char*) &tribes_players,	/* player_packet */
-    sizeof( tribes_players),	/* player_len */
+    sizeof(tribes_players),	/* player_len */
     (char*) &tribes_players,	/* rule_packet */
-    sizeof( tribes_players),	/* rule_len */
+    sizeof(tribes_players),	/* rule_len */
     NULL,			/* master_packet */
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_tribes_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_tribes_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_tribes_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_tribes_player_info,	/* display_xml_player_func */
@@ -1559,8 +1513,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_bfris_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_bfris_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_bfris_player_info,/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_bfris_player_info,	/* display_xml_player_func */
@@ -1586,7 +1540,7 @@ server_type builtin_types[] = {
     "gamedir",			/* game_rule */
     "KINGPIN",			/* template_var */
     (char*) &qw_serverstatus,	/* status_packet */
-    sizeof( qw_serverstatus),	/* status_len */
+    sizeof(qw_serverstatus),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -1595,8 +1549,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_q2_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_q2_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_q2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_q2_player_info,	/* display_xml_player_func */
@@ -1622,7 +1576,7 @@ server_type builtin_types[] = {
     "gamedir",			/* game_rule */
     "HERETIC2",			/* template_var */
     (char*) &qw_serverstatus,	/* status_packet */
-    sizeof( qw_serverstatus),	/* status_len */
+    sizeof(qw_serverstatus),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -1631,8 +1585,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_q2_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_q2_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_q2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_q2_player_info,	/* display_xml_player_func */
@@ -1658,7 +1612,7 @@ server_type builtin_types[] = {
     "gamedir",			/* game_rule */
     "SOLDIEROFFORTUNE",		/* template_var */
     (char*) &qw_serverstatus,	/* status_packet */
-    sizeof( qw_serverstatus),	/* status_len */
+    sizeof(qw_serverstatus),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -1667,8 +1621,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_q2_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_q2_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_q2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_q2_player_info,	/* display_xml_player_func */
@@ -1694,7 +1648,7 @@ server_type builtin_types[] = {
     "gametype",			/* game_rule */
     "GAMESPYPROTOCOL",		/* template_var */
     (char*) &unreal_serverstatus,	/* status_packet */
-    sizeof( unreal_serverstatus),	/* status_len */
+    sizeof(unreal_serverstatus),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -1703,8 +1657,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_unreal_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_unreal_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_unreal_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_unreal_player_info,	/* display_xml_player_func */
@@ -1730,17 +1684,17 @@ server_type builtin_types[] = {
     "game",			/* game_rule */
     "TRIBES2",			/* template_var */
     (char*) &tribes2_ping,	/* status_packet */
-    sizeof( tribes2_ping),	/* status_len */
+    sizeof(tribes2_ping),	/* status_len */
     (char*) &tribes2_info,	/* player_packet */
-    sizeof( tribes2_info),	/* player_len */
+    sizeof(tribes2_info),	/* player_len */
     (char*) NULL,		/* rule_packet */
     0,				/* rule_len */
     NULL,			/* master_packet */
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_tribes2_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_tribes2_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_tribes2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_tribes2_player_info,	/* display_xml_player_func */
@@ -1766,7 +1720,7 @@ server_type builtin_types[] = {
     "gametype",			/* game_rule */
     "DESCENT3",			/* template_var */
     (char*) &unreal_serverstatus,	/* status_packet */
-    sizeof( unreal_serverstatus),	/* status_len */
+    sizeof(unreal_serverstatus),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -1775,8 +1729,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_descent3_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_descent3_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_descent3_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_descent3_player_info,	/* display_xml_player_func */
@@ -1802,17 +1756,17 @@ server_type builtin_types[] = {
     "gametype",			/* game_rule */
     "DESCENT3",			/* template_var */
     (char*) &descent3_tcpipinfoquery,	/* status_packet */
-    sizeof( descent3_tcpipinfoquery),	/* status_len */
+    sizeof(descent3_tcpipinfoquery),	/* status_len */
     (char*) &descent3_playerquery,	/* player_packet */
-    sizeof( descent3_playerquery),	/* player_len */
+    sizeof(descent3_playerquery),	/* player_len */
     NULL,			/* rule_packet */
     0,				/* rule_len */
     NULL,			/* master_packet */
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_descent3_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_descent3_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_descent3_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_descent3_player_info,	/* display_xml_player_func */
@@ -1838,17 +1792,17 @@ server_type builtin_types[] = {
     "gametype",			/* game_rule */
     "DESCENT3",			/* template_var */
     (char*) &descent3_pxoinfoquery,	/* status_packet */
-    sizeof( descent3_pxoinfoquery),	/* status_len */
+    sizeof(descent3_pxoinfoquery),	/* status_len */
     (char*) &descent3_playerquery,	/* player_packet */
-    sizeof( descent3_playerquery),	/* player_len */
+    sizeof(descent3_playerquery),	/* player_len */
     NULL,			/* rule_packet */
     0,				/* rule_len */
     NULL,			/* master_packet */
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_descent3_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_descent3_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_descent3_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_descent3_player_info,	/* display_xml_player_func */
@@ -1874,7 +1828,7 @@ server_type builtin_types[] = {
     "gametype",			/* game_rule */
     "GHOSTRECON",			/* template_var */
     (char*) &ghostrecon_playerquery,	/* status_packet */
-    sizeof( ghostrecon_playerquery),	/* status_len */
+    sizeof(ghostrecon_playerquery),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -1883,8 +1837,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_ghostrecon_player_info,	/* display_player_func */
-    display_server_rules,		/* display_rule_func */
+    standard_display_ghostrecon_player_info,	/* display_player_func */
+    standard_display_server_rules,		/* display_rule_func */
     raw_display_ghostrecon_player_info,	/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_ghostrecon_player_info,	/* display_xml_player_func */
@@ -1910,7 +1864,7 @@ server_type builtin_types[] = {
     "gametype",				/* game_rule */
     "EYEPROTOCOL",		/* template_var */
     (char*) &eye_status_query,	/* status_packet */
-    sizeof( eye_status_query),	/* status_len */
+    sizeof(eye_status_query),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -1919,8 +1873,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_eye_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_eye_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_eye_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_eye_player_info,	/* display_xml_player_func */
@@ -1946,7 +1900,7 @@ server_type builtin_types[] = {
     "gametype",			/* game_rule */
     "GPS2PROTOCOL",		/* template_var */
     (char*) &gs2_status_query,	/* status_packet */
-    sizeof( gs2_status_query),	/* status_len */
+    sizeof(gs2_status_query),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -1955,8 +1909,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_gs2_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_gs2_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_gs2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_player_info,	/* display_xml_player_func */
@@ -1982,7 +1936,7 @@ server_type builtin_types[] = {
     "gametype",			/* game_rule */
     "AMERICASARMY",		/* template_var */
     (char*) &gs2_status_query,	/* status_packet */
-    sizeof( gs2_status_query),	/* status_len */
+    sizeof(gs2_status_query),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -1991,8 +1945,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_armyops_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_armyops_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_armyops_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_armyops_player_info,	/* display_xml_player_func */
@@ -2018,7 +1972,7 @@ server_type builtin_types[] = {
     "gametype",			/* game_rule */
     "RAVENSHIELD",			/* template_var */
     (char*)ravenshield_serverquery,			/* status_packet */
-    sizeof( ravenshield_serverquery ) - 1,	/* status_len */
+    sizeof(ravenshield_serverquery ) - 1,	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -2027,8 +1981,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_ravenshield_player_info,	/* display_player_func */
-    display_server_rules,		/* display_rule_func */
+    standard_display_ravenshield_player_info,	/* display_player_func */
+    standard_display_server_rules,		/* display_rule_func */
     raw_display_ravenshield_player_info,	/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_ravenshield_player_info,	/* display_xml_player_func */
@@ -2054,17 +2008,17 @@ server_type builtin_types[] = {
     "gametype",			/* game_rule */
     "SAVAGE",			/* template_var */
     (char*)savage_serverquery,			/* status_packet */
-    sizeof( savage_serverquery ) - 1,	/* status_len */
+    sizeof(savage_serverquery ) - 1,	/* status_len */
     (char*)savage_playerquery,			/* player_packet */
-    sizeof( savage_playerquery ) - 1,				/* player_len */
+    sizeof(savage_playerquery ) - 1,				/* player_len */
     NULL,			/* rule_packet */
     0,				/* rule_len */
     NULL,			/* master_packet */
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_savage_player_info,	/* display_player_func */
-    display_server_rules,		/* display_rule_func */
+    standard_display_savage_player_info,	/* display_player_func */
+    standard_display_server_rules,		/* display_rule_func */
     raw_display_savage_player_info,	/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_savage_player_info,	/* display_xml_player_func */
@@ -2090,7 +2044,7 @@ server_type builtin_types[] = {
     "gametype",			/* game_rule */
     "FARCRY",			/* template_var */
     (char*)farcry_serverquery,			/* status_packet */
-    sizeof( savage_serverquery ) - 1,	/* status_len */
+    sizeof(savage_serverquery ) - 1,	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -2099,8 +2053,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_farcry_player_info,	/* display_player_func */
-    display_server_rules,		/* display_rule_func */
+    standard_display_farcry_player_info,	/* display_player_func */
+    standard_display_server_rules,		/* display_rule_func */
     raw_display_farcry_player_info,	/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_farcry_player_info,	/* display_xml_player_func */
@@ -2135,10 +2089,10 @@ server_type builtin_types[] = {
     NULL,			/* rule_packet */
     0,				/* rule_len */
     (char*) &qw_masterquery,	/* master_packet */
-    sizeof( qw_masterquery),	/* master_len */
+    sizeof(qw_masterquery),	/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_qwmaster,		/* display_player_func */
+    standard_display_qwmaster,		/* display_player_func */
     NULL,	/* display_rule_func */
     NULL,	/* display_raw_player_func */
     NULL,	/* display_raw_rule_func */
@@ -2171,10 +2125,10 @@ server_type builtin_types[] = {
     NULL,			/* rule_packet */
     0,				/* rule_len */
     (char*) &hw_masterquery,	/* master_packet */
-    sizeof( hw_masterquery),	/* master_len */
+    sizeof(hw_masterquery),	/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_qwmaster,		/* display_player_func */
+    standard_display_qwmaster,		/* display_player_func */
     NULL,	/* display_rule_func */
     NULL,	/* display_raw_player_func */
     NULL,	/* display_raw_rule_func */
@@ -2207,10 +2161,10 @@ server_type builtin_types[] = {
     NULL,			/* rule_packet */
     0,				/* rule_len */
     q2_masterquery,		/* master_packet */
-    sizeof( q2_masterquery),	/* master_len */
+    sizeof(q2_masterquery),	/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_qwmaster,		/* display_player_func */
+    standard_display_qwmaster,		/* display_player_func */
     NULL,	/* display_rule_func */
     NULL,	/* display_raw_player_func */
     NULL,	/* display_raw_rule_func */
@@ -2246,7 +2200,7 @@ server_type builtin_types[] = {
     0,				/* master_len */
     q3_master_default_protocol,	/* master_protocol */
     q3_master_default_query,	/* master_query */
-    display_qwmaster,		/* display_player_func */
+    standard_display_qwmaster,		/* display_player_func */
     NULL,	/* display_rule_func */
     NULL,	/* display_raw_player_func */
     NULL,	/* display_raw_rule_func */
@@ -2282,7 +2236,7 @@ server_type builtin_types[] = {
     0,          /* master_len */
     NULL,	/* master_protocol */
     NULL,	/* master_query */
-    display_qwmaster,		/* display_player_func */
+    standard_display_qwmaster,		/* display_player_func */
     NULL,	/* display_rule_func */
     NULL,	/* display_raw_player_func */
     NULL,	/* display_raw_rule_func */
@@ -2318,7 +2272,7 @@ server_type builtin_types[] = {
     0,				/* master_len */
     rtcw_master_default_protocol,	/* master_protocol */
     q3_master_default_query,		/* master_query */
-    display_qwmaster,		/* display_player_func */
+    standard_display_qwmaster,		/* display_player_func */
     NULL,	/* display_rule_func */
     NULL,	/* display_raw_player_func */
     NULL,	/* display_raw_rule_func */
@@ -2354,7 +2308,7 @@ server_type builtin_types[] = {
     0,				/* master_len */
     stef_master_default_protocol,	/* master_protocol */
     q3_master_default_query,		/* master_query */
-    display_qwmaster,		/* display_player_func */
+    standard_display_qwmaster,		/* display_player_func */
     NULL,	/* display_rule_func */
     NULL,	/* display_raw_player_func */
     NULL,	/* display_raw_rule_func */
@@ -2390,7 +2344,7 @@ server_type builtin_types[] = {
     0,				/* master_len */
     jk3_master_default_protocol,	/* master_protocol */
     NULL,		/* master_query */
-    display_qwmaster,		/* display_player_func */
+    standard_display_qwmaster,		/* display_player_func */
     NULL,	/* display_rule_func */
     NULL,	/* display_raw_player_func */
     NULL,	/* display_raw_rule_func */
@@ -2423,10 +2377,10 @@ server_type builtin_types[] = {
     NULL,			/* rule_packet */
     0,				/* rule_len */
     (char*) &new_hl_masterquery_prefix,	/* master_packet */
-    sizeof( new_hl_masterquery_prefix),	/* master_len */
+    sizeof(new_hl_masterquery_prefix),	/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_qwmaster,		/* display_player_func */
+    standard_display_qwmaster,		/* display_player_func */
     NULL,	/* display_rule_func */
     NULL,	/* display_raw_player_func */
     NULL,	/* display_raw_rule_func */
@@ -2459,10 +2413,10 @@ server_type builtin_types[] = {
     NULL,			/* rule_packet */
     0,				/* rule_len */
     (char*) &tribes_masterquery,/* master_packet */
-    sizeof( tribes_masterquery),/* master_len */
+    sizeof(tribes_masterquery),/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_qwmaster,		/* display_player_func */
+    standard_display_qwmaster,		/* display_player_func */
     NULL,	/* display_rule_func */
     NULL,	/* display_raw_player_func */
     NULL,	/* display_raw_rule_func */
@@ -2489,16 +2443,16 @@ server_type builtin_types[] = {
     "",				/* game_rule */
     "GAMESPYMASTER",		/* template_var */
     (char*) &gamespy_master_request_prefix,	/* status_packet */
-    sizeof( gamespy_master_request_prefix)-1,	/* status_len */
+    sizeof(gamespy_master_request_prefix)-1,	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
     0,				/* rule_len */
     (char*) &gamespy_master_validate,/* master_packet */
-    sizeof( gamespy_master_validate)-1,/* master_len */
+    sizeof(gamespy_master_validate)-1,/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_qwmaster,		/* display_player_func */
+    standard_display_qwmaster,			/* display_player_func */
     NULL,	/* display_rule_func */
     NULL,	/* display_raw_player_func */
     NULL,	/* display_raw_rule_func */
@@ -2506,10 +2460,10 @@ server_type builtin_types[] = {
     NULL,	/* display_xml_rule_func */
     NULL,   /* display_json_player_func */
     NULL,   /* display_json_rule_func */
-    send_gamespy_master_request,/* status_query_func */
-    NULL,			/* rule_query_func */
-    NULL,			/* player_query_func */
-    deal_with_gamespy_master_response,/* packet_func */
+    send_gamespy_master_request,		/* status_query_func */
+    NULL,								/* rule_query_func */
+    NULL,								/* player_query_func */
+    deal_with_gamespy_master_response,	/* packet_func */
 },
 {
     /* TRIBES 2 MASTER */
@@ -2531,10 +2485,10 @@ server_type builtin_types[] = {
     NULL,			/* rule_packet */
     0,				/* rule_len */
     (char*) &tribes2_masterquery,/* master_packet */
-    sizeof( tribes2_masterquery),/* master_len */
+    sizeof(tribes2_masterquery),/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_qwmaster,		/* display_player_func */
+    standard_display_qwmaster,	/* display_player_func */
     NULL,	/* display_rule_func */
     NULL,	/* display_raw_player_func */
     NULL,	/* display_raw_rule_func */
@@ -2567,10 +2521,10 @@ server_type builtin_types[] = {
     NULL,			/* rule_packet */
     0,				/* rule_len */
     (char*)descent3_masterquery,	/* master_packet */
-    sizeof( descent3_masterquery),	/* master_len */
+    sizeof(descent3_masterquery),	/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_qwmaster,		/* display_player_func */
+    standard_display_qwmaster,		/* display_player_func */
     NULL,	/* display_rule_func */
     NULL,	/* display_raw_player_func */
     NULL,	/* display_raw_rule_func */
@@ -2606,7 +2560,7 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_qwmaster,		/* display_player_func */
+    standard_display_qwmaster,	/* display_player_func */
     NULL,	/* display_rule_func */
     NULL,	/* display_raw_player_func */
     NULL,	/* display_raw_rule_func */
@@ -2643,7 +2597,7 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,	/* master_protocol */
     NULL,	/* master_query */
-    display_qwmaster,		/* display_player_func */
+    standard_display_qwmaster,	/* display_player_func */
     NULL,	/* display_rule_func */
     NULL,	/* display_raw_player_func */
     NULL,	/* display_raw_rule_func */
@@ -2658,39 +2612,39 @@ server_type builtin_types[] = {
 },
 {
     /* HALFLIFE 2 */
-    A2S_SERVER,						/* id */
-    "A2S",							/* type_prefix */
-    "a2s",							/* type_string */
-    "-a2s",						/* type_option */
-    "Half-Life 2 new",					/* game_name */
-    0,								/* master */
-    HL2_DEFAULT_PORT,				/* default_port */
-    0,								/* port_offset */
-    TF_QUAKE3_NAMES,				/* flags */
+    A2S_SERVER,								/* id */
+    "A2S",									/* type_prefix */
+    "a2s",									/* type_string */
+    "-a2s",									/* type_option */
+    "Half-Life 2 new",						/* game_name */
+    0,										/* master */
+    HL2_DEFAULT_PORT,						/* default_port */
+    0,										/* port_offset */
+    TF_QUAKE3_NAMES,						/* flags */
     "gamedir",								/* game_rule */
-    "A2S",							/* template_var */
-    NULL,		/* status_packet */
-    0,		/* status_len */
-    NULL,		/* player_packet */
-    0,		/* player_len */
-    NULL,			/* rule_packet */
-    0,			/* rule_len */
-    NULL,							/* master_packet */
-    0,								/* master_len */
-    NULL,							/* master_protocol */
-    NULL,							/* master_query */
-    display_halflife_player_info,		/* display_player_func */
-    display_server_rules,			/* display_rule_func */
-    raw_display_halflife_player_info,	/* display_raw_player_func */
-    raw_display_server_rules,		/* display_raw_rule_func */
-    xml_display_halflife_player_info,	/* display_xml_player_func */
-    xml_display_server_rules,		/* display_xml_rule_func */
-    json_display_halflife_player_info,	/* display_json_player_func */
-    json_display_server_rules,			/* display_json_rule_func */
-    send_a2s_request_packet,		/* status_query_func */
-    send_a2s_rule_request_packet,	/* rule_query_func */
-    NULL,				/* player_query_func */
-    deal_with_a2s_packet,		/* packet_func */
+    "A2S",									/* template_var */
+    NULL,									/* status_packet */
+    0,										/* status_len */
+    NULL,									/* player_packet */
+    0,										/* player_len */
+    NULL,									/* rule_packet */
+    0,										/* rule_len */
+    NULL,									/* master_packet */
+    0,										/* master_len */
+    NULL,									/* master_protocol */
+    NULL,									/* master_query */
+    standard_display_halflife_player_info,	/* display_player_func */
+    standard_display_server_rules,			/* display_rule_func */
+    raw_display_halflife_player_info,		/* display_raw_player_func */
+    raw_display_server_rules,				/* display_raw_rule_func */
+    xml_display_halflife_player_info,		/* display_xml_player_func */
+    xml_display_server_rules,				/* display_xml_rule_func */
+    json_display_halflife_player_info,		/* display_json_player_func */
+    json_display_server_rules,				/* display_json_rule_func */
+    send_a2s_request_packet,				/* status_query_func */
+    send_a2s_rule_request_packet,			/* rule_query_func */
+    NULL,									/* player_query_func */
+    deal_with_a2s_packet,					/* packet_func */
 },
 {
     /* PARIAH */
@@ -2706,7 +2660,7 @@ server_type builtin_types[] = {
     "gametype",			/* game_rule */
     "UNREALTOURNAMENT2003",	/* template_var */
     (char*) &pariah_basicstatus,/* status_packet */
-    sizeof( pariah_basicstatus),/* status_len */
+    sizeof(pariah_basicstatus),/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     NULL,			/* rule_packet */
@@ -2715,8 +2669,8 @@ server_type builtin_types[] = {
     0,				/* master_len */
     NULL,			/* master_protocol */
     0,				/* master_query */
-    display_unreal_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_unreal_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_unreal_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_unreal_player_info,	/* display_xml_player_func */
@@ -2742,17 +2696,17 @@ server_type builtin_types[] = {
     "gametype",			/* game_rule */
     "GPS3PROTOCOL",		/* template_var */
     (char*) &gs3_status_query,	/* status_packet */
-    sizeof( gs3_status_query),	/* status_len */
+    sizeof(gs3_status_query),	/* status_len */
     (char*) &gs3_player_query,	/* player_packet */
-    sizeof( gs3_player_query),	/* player_len */
+    sizeof(gs3_player_query),	/* player_len */
     NULL,			/* rule_packet */
     0,				/* rule_len */
     NULL,			/* master_packet */
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_gs2_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_gs2_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_gs2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_player_info,	/* display_xml_player_func */
@@ -2787,8 +2741,8 @@ server_type builtin_types[] = {
     0,								/* master_len */
     NULL,							/* master_protocol */
     NULL,							/* master_query */
-    display_ts2_player_info,		/* display_player_func */
-    display_server_rules,			/* display_rule_func */
+    standard_display_ts2_player_info,		/* display_player_func */
+    standard_display_server_rules,		/* display_rule_func */
     raw_display_ts2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_ts2_player_info,	/* display_xml_player_func */
@@ -2814,7 +2768,7 @@ server_type builtin_types[] = {
     "fs_game",						/* game_rule */
     "QUAKE4",						/* template_var */
     (char*) &doom3_serverinfo,		/* status_packet */
-    sizeof( doom3_serverinfo),		/* status_len */
+    sizeof(doom3_serverinfo),		/* status_len */
     NULL,							/* player_packet */
     0,								/* player_len */
     NULL,							/* rule_packet */
@@ -2823,8 +2777,8 @@ server_type builtin_types[] = {
     0,								/* master_len */
     NULL,							/* master_protocol */
     NULL,							/* master_query */
-    display_doom3_player_info,		/* display_player_func */
-    display_server_rules,			/* display_rule_func */
+    standard_display_doom3_player_info,	/* display_player_func */
+    standard_display_server_rules,		/* display_rule_func */
     raw_display_doom3_player_info,	/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_doom3_player_info,	/* display_xml_player_func */
@@ -2859,7 +2813,7 @@ server_type builtin_types[] = {
     0,          /* master_len */
     NULL,	/* master_protocol */
     NULL,	/* master_query */
-    display_qwmaster,		/* display_player_func */
+    standard_display_qwmaster,		/* display_player_func */
     NULL,	/* display_rule_func */
     NULL,	/* display_raw_player_func */
     NULL,	/* display_raw_rule_func */
@@ -2886,17 +2840,17 @@ server_type builtin_types[] = {
     "gametype",			/* game_rule */
     "GPS4PROTOCOL",		/* template_var */
     (char*) &gs3_challenge,	/* status_packet */
-    sizeof( gs3_challenge),	/* status_len */
+    sizeof(gs3_challenge),	/* status_len */
     (char*) &gs3_challenge,	/* player_packet */
-    sizeof( gs3_challenge),	/* player_len */
+    sizeof(gs3_challenge),	/* player_len */
     NULL,			/* rule_packet */
     0,				/* rule_len */
     NULL,			/* master_packet */
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_gs2_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_gs2_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_gs2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_player_info,	/* display_xml_player_func */
@@ -2922,7 +2876,7 @@ server_type builtin_types[] = {
     "fs_game",						/* game_rule */
     "PREY",						/* template_var */
     (char*) &doom3_serverinfo,		/* status_packet */
-    sizeof( doom3_serverinfo),		/* status_len */
+    sizeof(doom3_serverinfo),		/* status_len */
     NULL,							/* player_packet */
     0,								/* player_len */
     NULL,							/* rule_packet */
@@ -2931,8 +2885,8 @@ server_type builtin_types[] = {
     0,								/* master_len */
     NULL,							/* master_protocol */
     NULL,							/* master_query */
-    display_doom3_player_info,		/* display_player_func */
-    display_server_rules,			/* display_rule_func */
+    standard_display_doom3_player_info,	/* display_player_func */
+    standard_display_server_rules,		/* display_rule_func */
     raw_display_doom3_player_info,	/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_doom3_player_info,	/* display_xml_player_func */
@@ -2967,8 +2921,8 @@ server_type builtin_types[] = {
     0,								/* master_len */
     NULL,							/* master_protocol */
     NULL,							/* master_query */
-    display_tm_player_info,			/* display_player_func */
-    display_server_rules,			/* display_rule_func */
+    standard_display_tm_player_info,		/* display_player_func */
+    standard_display_server_rules,		/* display_rule_func */
     raw_display_tm_player_info,		/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_tm_player_info,		/* display_xml_player_func */
@@ -2994,7 +2948,7 @@ server_type builtin_types[] = {
     "fs_game",						/* game_rule */
     "QUAKE4",						/* template_var */
     (char*) &doom3_serverinfo,		/* status_packet */
-    sizeof( doom3_serverinfo),		/* status_len */
+    sizeof(doom3_serverinfo),		/* status_len */
     NULL,							/* player_packet */
     0,								/* player_len */
     NULL,							/* rule_packet */
@@ -3003,8 +2957,8 @@ server_type builtin_types[] = {
     0,								/* master_len */
     NULL,							/* master_protocol */
     NULL,							/* master_query */
-    display_doom3_player_info,		/* display_player_func */
-    display_server_rules,			/* display_rule_func */
+    standard_display_doom3_player_info,	/* display_player_func */
+    standard_display_server_rules,		/* display_rule_func */
     raw_display_doom3_player_info,	/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_doom3_player_info,	/* display_xml_player_func */
@@ -3030,17 +2984,17 @@ server_type builtin_types[] = {
     "gametype",			/* game_rule */
     "HAZE",		/* template_var */
     (char*) &haze_status_query,	/* status_packet */
-    sizeof( haze_status_query),	/* status_len */
+    sizeof(haze_status_query),	/* status_len */
     (char*) &haze_player_query,	/* player_packet */
-    sizeof( haze_player_query),	/* player_len */
+    sizeof(haze_player_query),	/* player_len */
     NULL,			/* rule_packet */
     0,				/* rule_len */
     NULL,			/* master_packet */
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_gs2_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_gs2_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_gs2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_player_info,	/* display_xml_player_func */
@@ -3075,8 +3029,8 @@ server_type builtin_types[] = {
     0,								/* master_len */
     NULL,							/* master_protocol */
     NULL,							/* master_query */
-    display_wic_player_info,		/* display_player_func */
-    display_server_rules,			/* display_rule_func */
+    standard_display_wic_player_info,		/* display_player_func */
+    standard_display_server_rules,		/* display_rule_func */
     raw_display_wic_player_info,	/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_wic_player_info,	/* display_xml_player_func */
@@ -3102,17 +3056,17 @@ server_type builtin_types[] = {
     "",				/* game_rule */
     "OPENTTD",			/* template_var */
     (char*) &ottd_serverinfo,	/* status_packet */
-    sizeof( ottd_serverinfo),	/* status_len */
+    sizeof(ottd_serverinfo),	/* status_len */
     NULL,			/* player_packet */
     0,				/* player_len */
     (char*) &ottd_serverdetails,/* rule_packet */
-    sizeof( ottd_serverdetails),   /* rule_len */
+    sizeof(ottd_serverdetails),   /* rule_len */
     NULL,			/* master_packet */
     0,				/* master_len */
     NULL,			/* master_protocol */
     NULL,			/* master_query */
-    display_q2_player_info,	/* display_player_func */
-    display_server_rules,	/* display_rule_func */
+    standard_display_q2_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_q2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,	/* display_raw_rule_func */
     xml_display_player_info,/* display_xml_player_func */
@@ -3147,7 +3101,7 @@ server_type builtin_types[] = {
     sizeof(ottd_master_query),/* master_len */
     NULL,	/* master_protocol */
     NULL,	/* master_query */
-    display_qwmaster,		/* display_player_func */
+    standard_display_qwmaster,		/* display_player_func */
     NULL,	/* display_rule_func */
     NULL,	/* display_raw_player_func */
     NULL,	/* display_raw_rule_func */
@@ -3183,11 +3137,11 @@ server_type builtin_types[] = {
     0,								/* master_len */
     NULL,							/* master_protocol */
     NULL,							/* master_query */
-    display_fl_player_info,		/* display_player_func */
-    display_server_rules,			/* display_rule_func */
-    raw_display_fl_player_info,	/* display_raw_player_func */
+    standard_display_fl_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
+    raw_display_fl_player_info,		/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
-    xml_display_fl_player_info,	/* display_xml_player_func */
+    xml_display_fl_player_info,		/* display_xml_player_func */
     xml_display_server_rules,		/* display_xml_rule_func */
     json_display_fl_player_info,	/* display_json_player_func */
     json_display_server_rules,		/* display_json_rule_func */
@@ -3210,7 +3164,7 @@ server_type builtin_types[] = {
     "fs_game",						/* game_rule */
     "QUAKE4",						/* template_var */
     (char*) &doom3_serverinfo,		/* status_packet */
-    sizeof( doom3_serverinfo),		/* status_len */
+    sizeof(doom3_serverinfo),		/* status_len */
     NULL,							/* player_packet */
     0,								/* player_len */
     NULL,							/* rule_packet */
@@ -3219,8 +3173,8 @@ server_type builtin_types[] = {
     0,								/* master_len */
     NULL,							/* master_protocol */
     NULL,							/* master_query */
-    display_doom3_player_info,		/* display_player_func */
-    display_server_rules,			/* display_rule_func */
+    standard_display_doom3_player_info,		/* display_player_func */
+    standard_display_server_rules,			/* display_rule_func */
     raw_display_doom3_player_info,	/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_doom3_player_info,	/* display_xml_player_func */
@@ -3238,15 +3192,15 @@ server_type builtin_types[] = {
     "TEE",							/* type_prefix */
     "tee",							/* type_string */
     "-tee",							/* type_option */
-    "Teeworlds",		/* game_name */
+    "Teeworlds",					/* game_name */
     0,								/* master */
     35515,							/* default_port */
     0,								/* port_offset */
-    0,				/* flags */
-    "gametype",							/* game_rule */
+    0,								/* flags */
+    "gametype",						/* game_rule */
     "TEE",							/* template_var */
-    tee_serverstatus,						/* status_packet */
-    sizeof(tee_serverstatus),					/* status_len */
+    tee_serverstatus,				/* status_packet */
+    sizeof(tee_serverstatus),		/* status_len */
     NULL,							/* player_packet */
     0,								/* player_len */
     NULL,							/* rule_packet */
@@ -3255,15 +3209,15 @@ server_type builtin_types[] = {
     0,								/* master_len */
     NULL,							/* master_protocol */
     NULL,							/* master_query */
-    display_tee_player_info,			/* display_player_func */
-    display_server_rules,			/* display_rule_func */
-    raw_display_tee_player_info,		/* display_raw_player_func */
+    standard_display_tee_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
+    raw_display_tee_player_info,	/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
-    xml_display_tee_player_info,		/* display_xml_player_func */
+    xml_display_tee_player_info,	/* display_xml_player_func */
     xml_display_server_rules,		/* display_xml_rule_func */
     json_display_tee_player_info,	/* display_json_player_func */
     json_display_server_rules,		/* display_json_rule_func */
-    send_tee_request_packet,			/* status_query_func */
+    send_tee_request_packet,		/* status_query_func */
     NULL,							/* rule_query_func */
     NULL,							/* player_query_func */
     deal_with_tee_packet,			/* packet_func */
@@ -3291,8 +3245,8 @@ server_type builtin_types[] = {
     0,								/* master_len */
     NULL,							/* master_protocol */
     NULL,							/* master_query */
-    display_ts3_player_info,		/* display_player_func */
-    display_server_rules,			/* display_rule_func */
+    standard_display_ts3_player_info,	/* display_player_func */
+    standard_display_server_rules,	/* display_rule_func */
     raw_display_ts3_player_info,	/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_ts3_player_info,	/* display_xml_player_func */
@@ -3327,8 +3281,8 @@ server_type builtin_types[] = {
     0,								/* master_len */
     NULL,							/* master_protocol */
     NULL,							/* master_query */
-    display_bfbc2_player_info,		/* display_player_func */
-    display_server_rules,			/* display_rule_func */
+    standard_display_bfbc2_player_info,		/* display_player_func */
+    standard_display_server_rules,			/* display_rule_func */
     raw_display_bfbc2_player_info,	/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_bfbc2_player_info,	/* display_xml_player_func */
@@ -3363,8 +3317,8 @@ server_type builtin_types[] = {
     0,									/* master_len */
     NULL,								/* master_protocol */
     NULL,								/* master_query */
-    display_ventrilo_player_info,		/* display_player_func */
-    display_server_rules,				/* display_rule_func */
+    standard_display_ventrilo_player_info,		/* display_player_func */
+    standard_display_server_rules,				/* display_rule_func */
     raw_display_ventrilo_player_info,	/* display_raw_player_func */
     raw_display_server_rules,			/* display_raw_rule_func */
     xml_display_ventrilo_player_info,	/* display_xml_player_func */
@@ -3400,7 +3354,7 @@ server_type builtin_types[] = {
     NULL,								/* master_protocol */
     NULL,								/* master_query */
     NULL,								/* display_player_func */
-    display_server_rules,				/* display_rule_func */
+    standard_display_server_rules,		/* display_rule_func */
     NULL,								/* display_raw_player_func */
     raw_display_server_rules,			/* display_raw_rule_func */
     NULL,								/* display_xml_player_func */
@@ -3436,7 +3390,7 @@ server_type builtin_types[] = {
     NULL,							/* master_protocol */
     NULL,							/* master_query */
     NULL,							/* display_player_func */
-    display_server_rules,			/* display_rule_func */
+    standard_display_server_rules,			/* display_rule_func */
     NULL,							/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_player_info,		/* display_xml_player_func */
@@ -3472,7 +3426,7 @@ server_type builtin_types[] = {
     NULL,							/* master_protocol */
     NULL,							/* master_query */
     NULL,							/* display_player_func */
-    display_server_rules,			/* display_rule_func */
+    standard_display_server_rules,	/* display_rule_func */
     NULL,							/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_player_info,		/* display_xml_player_func */
@@ -3508,7 +3462,7 @@ server_type builtin_types[] = {
     NULL,							/* master_protocol */
     NULL,							/* master_query */
     NULL,							/* display_player_func */
-    display_server_rules,			/* display_rule_func */
+    standard_display_server_rules,			/* display_rule_func */
     NULL,							/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_player_info,		/* display_xml_player_func */
@@ -3544,7 +3498,7 @@ server_type builtin_types[] = {
     NULL,							/* master_protocol */
     NULL,							/* master_query */
     NULL,							/* display_player_func */
-    display_server_rules,			/* display_rule_func */
+    standard_display_server_rules,	/* display_rule_func */
     NULL,							/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_player_info,		/* display_xml_player_func */
@@ -3579,12 +3533,12 @@ server_type builtin_types[] = {
     0,								/* master_len */
     NULL,							/* master_protocol */
     NULL,							/* master_query */
-    display_starmade_player_info,	/* display_player_func */
-    display_server_rules,			/* display_rule_func */
-    raw_display_starmade_player_info,	/* display_raw_player_func */
-    raw_display_server_rules,		/* display_raw_rule_func */
-    xml_display_starmade_player_info,	/* display_xml_player_func */
-    xml_display_server_rules,		/* display_xml_rule_func */
+    standard_display_starmade_player_info,	/* display_player_func */
+    standard_display_server_rules,			/* display_rule_func */
+    raw_display_starmade_player_info,		/* display_raw_player_func */
+    raw_display_server_rules,				/* display_raw_rule_func */
+    xml_display_starmade_player_info,		/* display_xml_player_func */
+    xml_display_server_rules,				/* display_xml_rule_func */
     json_display_starmade_player_info,		/* display_json_player_func */
     json_display_server_rules,				/* display_json_rule_func */
     send_starmade_request_packet,	/* status_query_func */
@@ -3616,7 +3570,7 @@ server_type builtin_types[] = {
     NULL,							/* master_protocol */
     NULL,							/* master_query */
     NULL,							/* display_player_func */
-    display_server_rules,			/* display_rule_func */
+    standard_display_server_rules,	/* display_rule_func */
     NULL,							/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_player_info,		/* display_xml_player_func */
@@ -3652,7 +3606,7 @@ server_type builtin_types[] = {
     NULL,							/* master_protocol */
     NULL,							/* master_query */
     NULL,							/* display_player_func */
-    display_server_rules,			/* display_rule_func */
+    standard_display_server_rules,	/* display_rule_func */
     NULL,							/* display_raw_player_func */
     raw_display_server_rules,		/* display_raw_rule_func */
     xml_display_player_info,		/* display_xml_player_func */
@@ -3861,40 +3815,39 @@ HL master: send 'i', master responds with a small 'j' packet
 
 #define QW_REQUEST_LENGTH 20
 
-int is_default_rule( struct rule *rule);
-char *xform_name( char*, struct qserver *server);
-char *quake_color( int color);
-char *play_time( int seconds, int show_seconds);
-char *ping_time( int ms);
-char *get_qw_game( struct qserver *server);
+int is_default_rule(struct rule *rule);
+char *quake_color(int color);
+char *play_time(int seconds, int show_seconds);
+char *ping_time(int ms);
+char *get_qw_game(struct qserver *server);
 
 
 /*
  * Query status and packet handling functions
  */
 
-int cleanup_qserver( struct qserver *server, int force);
-void change_server_port( struct qserver *server, unsigned short port, int force );
+int cleanup_qserver(struct qserver *server, int force);
+void change_server_port(struct qserver *server, unsigned short port, int force );
 
-int server_info_packet( struct qserver *server, struct q_packet *pkt, int datalen );
-int player_info_packet( struct qserver *server, struct q_packet *pkt, int datalen );
-int rule_info_packet( struct qserver *server, struct q_packet *pkt, int datalen );
+int server_info_packet(struct qserver *server, struct q_packet *pkt, int datalen );
+int player_info_packet(struct qserver *server, struct q_packet *pkt, int datalen );
+int rule_info_packet(struct qserver *server, struct q_packet *pkt, int datalen );
 
-int time_delta( struct timeval *later, struct timeval *past);
-char * strherror( int h_err);
+int time_delta(struct timeval *later, struct timeval *past);
+char * strherror(int h_err);
 int connection_refused();
 int connection_would_block();
 int connection_reset();
 
-void add_file( char *filename);
-int add_qserver( char *arg, server_type* type, char *outfilename, char *query_arg);
-struct qserver* add_qserver_byaddr( unsigned int ipaddr, unsigned short port, server_type* type, int *new_server);
-void init_qserver( struct qserver *server, server_type* type);
-int bind_qserver( struct qserver *server);
+void add_file(char *filename);
+int add_qserver(char *arg, server_type* type, char *outfilename, char *query_arg);
+struct qserver* add_qserver_byaddr(unsigned int ipaddr, unsigned short port, server_type* type, int *new_server);
+void init_qserver(struct qserver *server, server_type* type);
+int bind_qserver(struct qserver *server);
 int bind_sockets();
 void send_packets();
-struct qserver * find_server_by_address( unsigned int ipaddr, unsigned short port);
-void add_server_to_hash( struct qserver *server);
+struct qserver * find_server_by_address(unsigned int ipaddr, unsigned short port);
+void add_server_to_hash(struct qserver *server);
 
 #define NO_FLAGS 0
 #define NO_VALUE_COPY 1
@@ -3903,52 +3856,28 @@ void add_server_to_hash( struct qserver *server);
 #define COMBINE_VALUES 8
 #define OVERWITE_DUPLICATES 16
 
-struct player* get_player_by_number( struct qserver *server, int player_number );
-struct rule* add_rule( struct qserver *server, char *key, char *value,	int flags) ;
-struct player* add_player( struct qserver *server, int player_number );
-struct info* player_add_info( struct player *player, char *key, char *value, int flags );
-void players_set_teamname( struct qserver *server, int teamid, char *teamname );
-
-
-/*
- * Output template stuff
- */
-
-int read_qserver_template( char *filename);
-int read_rule_template( char *filename);
-int read_header_template( char *filename);
-int read_trailer_template( char *filename);
-int read_player_template( char *filename);
-int have_server_template();
-int have_header_template();
-int have_trailer_template();
-
-void template_display_server( struct qserver *server);
-void template_display_header();
-void template_display_trailer();
-void template_display_players( struct qserver *server);
-void template_display_player( struct qserver *server, struct player *player);
-void template_display_rules( struct qserver *server);
-void template_display_rule( struct qserver *server, struct rule *rule);
-
-
+struct player* get_player_by_number(struct qserver *server, int player_number );
+struct rule* add_rule(struct qserver *server, char *key, char *value,	int flags) ;
+struct player* add_player(struct qserver *server, int player_number );
+struct info* player_add_info(struct player *player, char *key, char *value, int flags );
+void players_set_teamname(struct qserver *server, int teamid, char *teamname );
 
 /*
  * Host cache stuff
  */
 
-int hcache_open( char *filename, int update);
-void hcache_write( char *filename);
+int hcache_open(char *filename, int update);
+void hcache_write(char *filename);
 void hcache_invalidate();
 void hcache_validate();
-unsigned long hcache_lookup_hostname( char *hostname);
-char * hcache_lookup_ipaddr( unsigned long ipaddr);
-void hcache_write_file( char *filename);
+unsigned long hcache_lookup_hostname(char *hostname);
+char * hcache_lookup_ipaddr(unsigned long ipaddr);
+void hcache_write_file(char *filename);
 void hcache_update_file();
 
-unsigned int swap_long_from_little( void *l);
-unsigned short swap_short_from_little( void *l);
-float swap_float_from_little( void *f);
+unsigned int swap_long_from_little(void *l);
+unsigned short swap_short_from_little(void *l);
+float swap_float_from_little(void *f);
 
 /** \brief write four bytes in little endian order */
 void put_long_little(unsigned val, char* buf);
